### 2024-06-26 b64a96d
add JOSS and arXiv badge
