### 2024-06-27 3f3c386
working on pixelated potential test
