+++ src/caustics/__init__.py
@@ -13,6 +13,7 @@ from .lenses import (
     EPL,
     ExternalShear,
     PixelatedConvergence,
+    PixelatedPotential,
     Multiplane,
     NFW,
     Point,
@@ -44,6 +45,7 @@ __all__ = [
     "EPL",
     "ExternalShear",
     "PixelatedConvergence",
+    "PixelatedPotential",
     "Multiplane",
     "NFW",
     "Point",
