+++ src/caustics/lenses/__init__.py
@@ -2,6 +2,7 @@ from .base import ThinLens, ThickLens
 from .epl import EPL
 from .external_shear import ExternalShear
 from .pixelated_convergence import PixelatedConvergence
+from .pixelated_potential import PixelatedPotential
 from .nfw import NFW
 from .point import Point
 from .pseudo_jaffe import PseudoJaffe
@@ -19,6 +20,7 @@ __all__ = [
     "EPL",
     "ExternalShear",
     "PixelatedConvergence",
+    "PixelatedPotential",
     "Multiplane",
     "NFW",
     "Point",
