### 2024-06-27 84c6105
All tests are now passing, missed a coupled Lens_Source in the registery
