+++ src/caustics/sims/__init__.py
@@ -1,4 +1,4 @@
 from .lens_source import LensSource
 from .simulator import Simulator
 
-__all__ = ["Lens_Source", "Simulator"]
+__all__ = ["LensSource", "Simulator"]
