+++ .pre-commit-config.yaml
@@ -70,11 +70,20 @@ repos:
       - id: codespell
         args: ["--write-changes", "--ignore-words", ".codespell-whitelist"]
 
-  - repo: https://github.com/kynan/nbstripout
-    rev: 0.7.1
+  #- repo: https://github.com/kynan/nbstripout
+  #rev: 0.7.1
+  #hooks:
+  #- id: nbstripout
+  #args: [--extra-keys=metadata.kernelspec metadata.language_info.version]
+
+  - repo: local
     hooks:
-      - id: nbstripout
-        args: [--extra-keys=metadata.kernelspec metadata.language_info.version]
+      - id: strip-output-keep-html
+        name: Strip Jupyter Notebook output but keep HTML
+        entry: python strip_output_keep_html.py
+        language: python
+        additional_dependencies: [nbformat]
+        files: \.ipynb$
 
   - repo: local
     hooks:
