+++ docs/source/_toc.yml
@@ -11,7 +11,7 @@ chapters:
     sections:
       - file: basicindex
         sections:
-          - file: tutorials/BasicIntroduction
+          - file: tutorials/Introduction
           - file: tutorials/BasicIntroduction_yaml
           - file: tutorials/BasicIntroduction_oop
           - file: tutorials/BasicIntroduction_func
