### 2024-06-27 a982cf4
Added assets to run the notebook, cleaned up some typos and errors
