### 2024-06-28 08d1b67
Refactored PixelatedConvergence so n_pix is no longer required at initialization. It's inferred from shape or the provided image
