### 2024-06-28 eb57842
Removed most of the get_
