+++ src/caustics/lenses/utils.py
@@ -5,10 +5,10 @@ from torch import Tensor
 
 from ..utils import vmap_n
 
-__all__ = ("get_pix_jacobian", "get_pix_magnification", "get_magnification")
+__all__ = ("pixel_jacobian", "pixel_magnification", "magnification")
 
 
-def get_pix_jacobian(
+def pixel_jacobian(
     raytrace, x, y, z_s
 ) -> Tuple[Tuple[Tensor, Tensor], Tuple[Tensor, Tensor]]:
     """Computes the Jacobian matrix of the partial derivatives of the
@@ -47,7 +47,7 @@ def get_pix_jacobian(
     return jac
 
 
-def get_pix_magnification(raytrace, x, y, z_s) -> Tensor:
+def pixel_magnification(raytrace, x, y, z_s) -> Tensor:
     """
     Computes the magnification at a single point on the lensing plane.
     The magnification is derived from the determinant
@@ -81,11 +81,11 @@ def get_pix_magnification(raytrace, x, y, z_s) -> Tensor:
         *Unit: unitless*
 
     """
-    jac = get_pix_jacobian(raytrace, x, y, z_s)
+    jac = pixel_jacobian(raytrace, x, y, z_s)
     return 1 / (jac[0][0] * jac[1][1] - jac[0][1] * jac[1][0]).abs()  # fmt: skip
 
 
-def get_magnification(raytrace, x, y, z_s) -> Tensor:
+def magnification(raytrace, x, y, z_s) -> Tensor:
     """
     Computes the magnification over a grid on the lensing plane.
     This is done by calling `get_pix_magnification`
@@ -119,4 +119,4 @@ def get_magnification(raytrace, x, y, z_s) -> Tensor:
         *Unit: unitless*
 
     """
-    return vmap_n(get_pix_magnification, 2, (None, 0, 0, None))(raytrace, x, y, z_s)
+    return vmap_n(pixel_magnification, 2, (None, 0, 0, None))(raytrace, x, y, z_s)
