### 2024-06-28 493c721
rearrange caustics intro tutorials
