+++ docs/source/interfaceindex.rst
@@ -1,5 +1,5 @@
-Basic Introduction
-==================
+The three interfaces to Caustics
+================================
 
 There are three user interfaces to Caustics: configuration file, object-oriented, and functional.
 Here we have built the same tutorial three times over so you can see how the three interfaces compare.
