+++ docs/source/_toc.yml
@@ -9,12 +9,12 @@ chapters:
   - file: websim
   - file: tutorials/index
     sections:
-      - file: basicindex
+      - file: tutorials/Introduction
+      - file: interfaceindex
         sections:
-          - file: tutorials/Introduction
-          - file: tutorials/BasicIntroduction_yaml
-          - file: tutorials/BasicIntroduction_oop
-          - file: tutorials/BasicIntroduction_func
+          - file: tutorials/InterfaceIntroduction_yaml
+          - file: tutorials/InterfaceIntroduction_oop
+          - file: tutorials/InterfaceIntroduction_func
       - file: tutorials/LensZoo
       - file: tutorials/VisualizeCaustics
       - file: tutorials/MultiplaneDemo
