### 2024-06-28 aec8402
Pixelated potential now passes checks
