### 2024-06-28 7d01249
add enclosed mass tests
