+++ src/caustics/lenses/enclosed_mass.py
@@ -2,23 +2,30 @@
 from typing import Optional, Union, Annotated, Callable
 
 from torch import Tensor
-import torch
 
-from ..utils import translate_rotate
 from .base import ThinLens, CosmologyType, NameType, ZLType
+from ..parametrized import unpack
 from ..packed import Packed
-from ..constants import G_over_c2
+from .func import physical_deflection_angle_enclosed_mass, convergence_enclosed_mass
 
-__all__ = ("Enclosed_Mass",)
+__all__ = ("EnclosedMass",)
 
 
-class Enclosed_Mass(ThinLens):
+class EnclosedMass(ThinLens):
     """
     A class for representing a lens with an enclosed mass profile. This generic
     lens profile can represent any lens with a mass distribution that can be
     described by a function that returns the enclosed mass as a function of
-    radius. Note that this will be an axisymmetric lens.
+    radius.
     """
+
+    _null_params = {
+        "x0": 0.0,
+        "y0": 0.0,
+        "q": 1.0,
+        "phi": 0.0,
+        "p": 1.0,
+    }
 
     def __init__(
         self,
@@ -30,6 +37,12 @@ class Enclosed_Mass(ThinLens):
         ] = None,
         y0: Annotated[
             Optional[Union[Tensor, float]], "The y-coordinate of the lens center", True
+        ] = None,
+        q: Annotated[
+            Optional[Union[Tensor, float]], "The axis ratio of the lens", True
+        ] = None,
+        phi: Annotated[
+            Optional[Union[Tensor, float]], "The position angle of the lens", True
         ] = None,
         p: Annotated[
             Optional[Union[Tensor, list[float]]],
@@ -73,6 +86,16 @@ class Enclosed_Mass(ThinLens):
 
             *Unit: arcsec*
 
+        q : float or Tensor, optional
+            The axis ratio of the lens. ratio of semi-minor to semi-major axis (b/a).
+
+            *Unit: unitless*
+
+        phi : float or Tensor, optional
+            The position angle of the lens.
+
+            *Unit: radians*
+
         p : list[float] or Tensor, optional
             The parameters for the enclosed mass function.
 
@@ -88,10 +111,13 @@ class Enclosed_Mass(ThinLens):
 
         self.add_param("x0", x0)
         self.add_param("y0", y0)
+        self.add_param("q", q)
+        self.add_param("phi", phi)
         self.add_param("p", p)
 
         self.s = s
 
+    @unpack
     def physical_deflection_angle(
         self,
         x: Tensor,
@@ -102,6 +128,8 @@ class Enclosed_Mass(ThinLens):
         z_l: Optional[Tensor] = None,
         x0: Optional[Tensor] = None,
         y0: Optional[Tensor] = None,
+        q: Optional[Tensor] = None,
+        phi: Optional[Tensor] = None,
         p: Optional[Tensor] = None,
         **kwargs,
     ) -> tuple[Tensor, Tensor]:
@@ -131,6 +159,16 @@ class Enclosed_Mass(ThinLens):
 
             *Unit: arcsec*
 
+        q : float or Tensor, optional
+            The axis ratio of the lens. ratio of semi-minor to semi-major axis (b/a).
+
+            *Unit: unitless*
+
+        phi : float or Tensor, optional
+            The position angle of the lens.
+
+            *Unit: radians*
+
         p : list[float] or Tensor, optional
             The parameters for the enclosed mass function.
 
@@ -142,12 +180,11 @@ class Enclosed_Mass(ThinLens):
 
             *Unit: arcsec*
         """
-        x, y = translate_rotate(x, y, x0, y0)
-        r = (x**2 + y**2).sqrt() + self.s
[…]
