+++ src/caustics/lenses/__init__.py
@@ -12,6 +12,7 @@ from .singleplane import SinglePlane
 from .mass_sheet import MassSheet
 from .tnfw import TNFW
 from .multiplane import Multiplane
+from .enclosed_mass import EnclosedMass
 
 
 __all__ = [
@@ -30,4 +31,5 @@ __all__ = [
     "SinglePlane",
     "MassSheet",
     "TNFW",
+    "EnclosedMass",
 ]
