### 2024-06-28 2b97f57
refactor: Fix magnification input shape to allow any shape (#220)
* fix magnification to take any shape

* test for magnification

* reshape magnification works for all cases now
