### 2024-06-28 471fb85
updated get_graph
