### 2024-06-28 31e66a8
refactor: Interface cleanup (#227)
* Refactored PixelatedConvergence so n_pix is no longer required at initialization. It's inferred from shape or the provided image

* Removed most of the get_

* Updated Introduction to reflect PixelatedConvergence change

* updated get_graph

* Refactored Lens_Source to LensSource in notebook and comments
