+++ tests/utils/models.py
@@ -33,9 +33,8 @@ def setup_complex_multiplane_yaml():
     res = 0.05
     upsample_factor = 2
     fov = res * n_pix
-    thx, thy = caustics.utils.get_meshgrid(
+    thx, thy = caustics.utils.meshgrid(
         res / upsample_factor,
-        upsample_factor * n_pix,
         upsample_factor * n_pix,
         dtype=torch.float32,
     )
