### 2024-06-28 c19621b
feat: Lens converter classes pixelated potential and enclosed mass (#203)
* Adding enclosed mass lens

* Adding pixelized potential, enclosed mass ellipse

* format

* working on pixelated potential test

* pix potential runs but is wrong

* Pixelated potential now passes checks

* add enclosed mass tests

* update meshgrid name
