+++ src/caustics/utils.py
@@ -1,6 +1,6 @@
-# mypy: disable-error-code="misc"
+# mypy: disable-error-code="misc", disable-error-code="attr-defined"
 from math import pi
-from typing import Callable, Optional, Tuple, Union, Any
+from typing import Callable, Optional, Tuple, Union, Any, Literal
 from importlib import import_module
 from functools import partial, lru_cache
 
@@ -318,7 +318,7 @@ def quad(
     pixelscale: float,
     X: Tensor,
     Y: Tensor,
-    args: Optional[Tuple] = None,
+    args: Tuple = (),
     quad_level: int = 3,
 ):
     """
@@ -417,7 +417,12 @@ def _h_poly(t):
     return A @ tt
 
 
-def interp1d(x: Tensor, y: Tensor, xs: Tensor, extend: str = "extrapolate") -> Tensor:
+def interp1d(
+    x: Tensor,
+    y: Tensor,
+    xs: Tensor,
+    extend: Literal["extrapolate", "const", "linear"] = "extrapolate",
+) -> Tensor:
     """Compute the 1D cubic spline interpolation for the given data points
     using PyTorch.
 
@@ -461,7 +466,7 @@ def interp2d(
     im: Tensor,
     x: Tensor,
     y: Tensor,
-    method: str = "linear",
+    method: Literal["linear", "nearest"] = "linear",
     padding_mode: str = "zeros",
 ) -> Tensor:
     """
@@ -552,8 +557,8 @@ def interp3d(
     x: Tensor,
     y: Tensor,
     t: Tensor,
-    method: str = "linear",
-    padding_mode: str = "zeros",
+    method: Literal["linear", "nearest"] = "linear",
+    padding_mode: Literal["zeros", "extrapolate"] = "zeros",
 ) -> Tensor:
     """
     Interpolates a 3D image at specified coordinates.
@@ -597,10 +602,7 @@ def interp3d(
     """
     if cu.ndim != 3:
         raise ValueError(f"im must be 3D (received {cu.ndim}D tensor)")
-    if x.ndim > 1:
-        raise ValueError(f"x must be 0 or 1D (received {x.ndim}D tensor)")
-    if y.ndim > 1:
-        raise ValueError(f"y must be 0 or 1D (received {y.ndim}D tensor)")
+
     if t.ndim > 1:
         raise ValueError(f"t must be 0 or 1D (received {t.ndim}D tensor)")
     if padding_mode not in ["extrapolate", "zeros"]:
@@ -662,6 +664,253 @@ def interp3d(
         result = torch.where(idxs_out_of_bounds, torch.zeros_like(result), result)
 
     return result
+
+
+# Bicubic interpolation coefficients
+# These are the coefficients for the bicubic interpolation kernel.
+# To quote numerical recipes:
+#     The formulas that obtain the c’s from the function and derivative values
+#     are just a complicated linear transformation, with coefficients which,
+#     having been determined once in the mists of numerical history, can be
+#     tabulated and forgotten
+BC = (
+    (1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
+    (0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0),
+    (-3, 0, 0, 3, 0, 0, 0, 0, -2, 0, 0, -1, 0, 0, 0, 0),
+    (2, 0, 0, -2, 0, 0, 0, 0, 1, 0, 0, 1, 0, 0, 0, 0),
[…]
