+++ src/caustics/lenses/pixelated_convergence.py
@@ -401,11 +401,6 @@ class PixelatedConvergence(ThinLens):
 
             *Unit: unitless*
 
-        Raises
-        ------
-        NotImplementedError
-            This method is not implemented.
-
         """
         return interp2d(
             convergence_map,
