### 2024-06-29 535764b
remove caustics data
