### 2024-07-01 62dde5d
ci: fix codecov token (#228)
* Attempt 1 to fix codecov token

* remove caustics data
