+++ src/caustics/light/__init__.py
@@ -1,7 +1,6 @@
 from .base import Source
 from .pixelated import Pixelated
-from .probes import PROBESDataset
 from .sersic import Sersic
 from .pixelated_time import PixelatedTime
 
-__all__ = ["Source", "Pixelated", "PixelatedTime", "PROBESDataset", "Sersic"]
+__all__ = ["Source", "Pixelated", "PixelatedTime", "Sersic"]
