+++ .github/workflows/ci.yml
@@ -71,4 +71,12 @@ jobs:
           pytest -vvv --cov=${{ env.PROJECT_NAME }} --cov-report=xml --cov-report=term tests/
 
       - name: Upload coverage reports to Codecov with GitHub Action
+        if:
+          ${{ matrix.python-version == '3.10' && matrix.os == 'ubuntu-latest'}}
         uses: codecov/codecov-action@v4
+        with:
+          token: ${{ secrets.CODECOV_TOKEN }}
+          files: ./coverage.xml
+          flags: unittests
+          name: codecov-umbrella
+          fail_ci_if_error: true
