### 2024-07-03 7c9840d
move time delay factor to func
