+++ src/caustics/lenses/base.py
@@ -7,7 +7,6 @@ import warnings
 import torch
 from torch import Tensor
 
-from ..constants import arcsec_to_rad, c_Mpc_s
 from ..cosmology import Cosmology
 from ..parametrized import Parametrized, unpack
 from .utils import magnification
@@ -1006,15 +1005,15 @@ class ThinLens(Lens):
         ax, ay = self.reduced_deflection_angle(x, y, z_s, params, **kwargs)
         return x - ax, y - ay
 
-    @staticmethod
-    def _arcsec2_to_time(z_l, z_s, cosmology, params):
+    def _arcsec2_to_time(self, z_l, z_s, params):
         """
-        This method is used by :func:`caustics.lenses.ThinLens.time_delay` to convert arcsec^2 to seconds in the context of gravitational time delays.
+        This method is used by :func:`caustics.lenses.ThinLens.time_delay` to
+        convert arcsec^2 to seconds in the context of gravitational time delays.
         """
-        d_l = cosmology.angular_diameter_distance(z_l, params)
-        d_s = cosmology.angular_diameter_distance(z_s, params)
-        d_ls = cosmology.angular_diameter_distance_z1z2(z_l, z_s, params)
-        return (1 + z_l) / c_Mpc_s * d_s * d_l / d_ls * arcsec_to_rad**2
+        d_l = self.cosmology.angular_diameter_distance(z_l, params)
+        d_s = self.cosmology.angular_diameter_distance(z_s, params)
+        d_ls = self.cosmology.angular_diameter_distance_z1z2(z_l, z_s, params)
+        return func.td_arcsec2_to_time(d_l, d_s, d_ls, z_l)
 
     @unpack
     def time_delay(
@@ -1090,13 +1089,13 @@ class ThinLens(Lens):
 
         if shapiro_time_delay:
             potential = self.potential(x, y, z_s, params)
-            TD -= potential
+            TD = TD - potential
         if geometric_time_delay:
             ax, ay = self.physical_deflection_angle(x, y, z_s, params)
             fp = 0.5 * (ax**2 + ay**2)
-            TD += fp
+            TD = TD + fp
 
-        factor = self._arcsec2_to_time(z_l, z_s, self.cosmology, params)
+        factor = self._arcsec2_to_time(z_l, z_s, params)
 
         return factor * TD
 
