+++ src/caustics/func.py
@@ -2,6 +2,7 @@ from .lenses.func import (
     forward_raytrace,
     physical_from_reduced_deflection_angle,
     reduced_from_physical_deflection_angle,
+    time_delay_arcsec2_to_time,
     reduced_deflection_angle_sie,
     potential_sie,
     convergence_sie,
@@ -53,6 +54,7 @@ __all__ = (
     "forward_raytrace",
     "physical_from_reduced_deflection_angle",
     "reduced_from_physical_deflection_angle",
+    "time_delay_arcsec2_to_time",
     "reduced_deflection_angle_sie",
     "potential_sie",
     "convergence_sie",
