+++ src/caustics/lenses/func/__init__.py
@@ -2,7 +2,7 @@ from .base import (
     forward_raytrace,
     physical_from_reduced_deflection_angle,
     reduced_from_physical_deflection_angle,
-    td_arcsec2_to_time,
+    time_delay_arcsec2_to_time,
 )
 from .sie import reduced_deflection_angle_sie, potential_sie, convergence_sie
 from .point import reduced_deflection_angle_point, potential_point, convergence_point
@@ -62,7 +62,7 @@ __all__ = (
     "forward_raytrace",
     "physical_from_reduced_deflection_angle",
     "reduced_from_physical_deflection_angle",
-    "td_arcsec2_to_time",
+    "time_delay_arcsec2_to_time",
     "reduced_deflection_angle_sie",
     "potential_sie",
     "convergence_sie",
