### 2024-07-05 31e5140
time to days in func time delay function
