+++ src/caustics/func.py
@@ -2,7 +2,7 @@ from .lenses.func import (
     forward_raytrace,
     physical_from_reduced_deflection_angle,
     reduced_from_physical_deflection_angle,
-    time_delay_arcsec2_to_time,
+    time_delay_arcsec2_to_days,
     reduced_deflection_angle_sie,
     potential_sie,
     convergence_sie,
@@ -54,7 +54,7 @@ __all__ = (
     "forward_raytrace",
     "physical_from_reduced_deflection_angle",
     "reduced_from_physical_deflection_angle",
-    "time_delay_arcsec2_to_time",
+    "time_delay_arcsec2_to_days",
     "reduced_deflection_angle_sie",
     "potential_sie",
     "convergence_sie",
