### 2024-07-05 9cea524
have multiplane use func time delay factor
