+++ docs/source/citation.rst
@@ -3,4 +3,23 @@
 Citation
 ========
 
-Paper coming soon! We will put all the citation information here when its ready.
+Paper is submitted to JOSS, in the meantime please cite the `arXiv version<https://arxiv.org/abs/2406.15542>`_`:
+
+.. code-block:: bibtex
+
+    @ARTICLE{2024arXiv240615542S,
+        author = {{Stone}, Connor and {Adam}, Alexandre and {Coogan}, Adam and {Yantovski-Barth}, M.~J. and {Filipp}, Andreas and {Setiawan}, Landung and {Core}, Cordero and {Legin}, Ronan and {Wilson}, Charles and {Missael Barco}, Gabriel and {Hezaveh}, Yashar and {Perreault-Levasseur}, Laurence},
+            title = "{Caustics: A Python Package for Accelerated Strong Gravitational Lensing Simulations}",
+        journal = {arXiv e-prints},
+        keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
+            year = 2024,
+            month = jun,
+            eid = {arXiv:2406.15542},
+            pages = {arXiv:2406.15542},
+            doi = {10.48550/arXiv.2406.15542},
+    archivePrefix = {arXiv},
+        eprint = {2406.15542},
+    primaryClass = {astro-ph.IM},
+        adsurl = {https://ui.adsabs.harvard.edu/abs/2024arXiv240615542S},
+        adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+    }
