### 2024-07-16 6ed06fa
try to make window inline
