+++ docs/source/_toc.yml
@@ -21,7 +21,6 @@ chapters:
       - file: tutorials/InvertLensEquation
       - file: tutorials/Parameters
       - file: tutorials/Simulators
-      - file: tutorials/Playground
   - file: examples/index
     sections:
       - file: examples/Example_ImageFit_LM
