+++ docs/source/websim.rst
@@ -7,9 +7,12 @@ Pro tip: check out the "Pixelated" source to lens any image you want!
 
 `Launch the simulator! <https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/>`_
 
+````{div} full-width
 .. raw:: html
 
-   <iframe src="https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/?embed=true" width="100%" height="800"></iframe>
+   <iframe src="https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/?embed=true" width="100%" height="80vh"></iframe>
+
+````
 
 For frequent simulator users (e.g., if you plan on exploring the parameter space of a lens), we recommend installing the simulator locally and running it in your browser. Follow the steps below:
 
