+++ docs/source/websim.md
@@ -6,9 +6,7 @@ having to install anything.
 
 Pro tip: check out the "Pixelated" source to lens any image you want!
 
-```{div} full-width
 <iframe src="https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/?embed=true" width="100%" height="80vh"></iframe>
-```
 
 For frequent simulator users (e.g., if you plan on exploring the parameter space
 of a lens), we recommend installing the simulator locally and running it in your
