+++ docs/source/websim.md
@@ -6,8 +6,8 @@ having to install anything.
 
 Pro tip: check out the "Pixelated" source to lens any image you want!
 
-<div style="position: relative; width: 100%; padding-top: 56.25%;">
-    <iframe src="https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/?embed=true" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%;"></iframe>
+<div style="display: flex; height: 500px;">
+    <iframe src="https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/?embed=true" style="flex-grow: 1; width: 100%;"></iframe>
 </div>
 
 <!-- <iframe src="https://ciela-institute-caustics-webapp-guistreamlit-app-yanhhm.streamlit.app/?embed=true" width="100%" height="80vh"></iframe> -->
