### 2024-07-19 4eeca03
Put logo gif in main docs page
