+++ docs/source/intro.md
@@ -1,6 +1,6 @@
 # Welcome to Caustics’ documentation!
 
-[Logo GIF](../../media/caustics_logo.gif)
+![Logo GIF](../../media/caustics_logo.gif)
 
 The lensing pipeline of the future: GPU-accelerated,
 automatically-differentiable, highly modular and extensible.
