+++ src/caustics/lenses/multiplane.py
@@ -1,10 +1,9 @@
-from operator import itemgetter
-from typing import Optional
+from typing import Optional, Union
 
 import torch
 from torch import Tensor
 
-from ..constants import arcsec_to_rad, rad_to_arcsec, c_Mpc_s, days_to_seconds
+from ..constants import arcsec_to_rad, c_Mpc_s, seconds_to_days
 from .base import ThickLens, NameType, CosmologyType, LensesType
 from ..parametrized import unpack
 from ..packed import Packed
@@ -70,75 +69,70 @@ class Multiplane(ThickLens):
         z_s: Tensor,
         *args,
         params: Optional["Packed"] = None,
-        shapiro_time_delay: bool = True,
-        geometric_time_delay: bool = True,
-        ray_coords: bool = True,
+        return_shapiro_time_delay: bool = False,
+        return_geometric_time_delay: bool = False,
+        return_time_delay: bool = False,
+        return_coordinates: bool = True,
         **kwargs,
-    ):
-        # Collect lens redshifts and ensure proper order
+    ) -> Union[tuple[Tensor], Tensor]:
+        # Order resdhift
         z_ls = self.get_z_ls(params)
-        lens_planes = [i for i, _ in sorted(enumerate(z_ls), key=itemgetter(1))]
-        D_s = self.cosmology.transverse_comoving_distance(z_s, params)
+        planes = torch.argsort(torch.stack(z_ls))
 
-        # Compute physical position on first lens plane
-        D = self.cosmology.transverse_comoving_distance(z_ls[lens_planes[0]], params)
-        X, Y = x * arcsec_to_rad * D, y * arcsec_to_rad * D  # fmt: skip
+        theta_x = x
+        theta_y = y
 
-        # Initial angles are observation angles
-        theta_x, theta_y = x, y
+        shapiro_tau = torch.zeros_like(x)
+        geometric_tau = torch.zeros_like(x)
 
-        # Store the time delays
-        if shapiro_time_delay or geometric_time_delay:
-            TD = torch.zeros_like(x)
-
-        for i in lens_planes:
-            z_next = z_ls[i + 1] if i != lens_planes[-1] else z_s
-            # Compute deflection angle at current ray positions
-            D_l = self.cosmology.transverse_comoving_distance(z_ls[i], params)
-            D = self.cosmology.transverse_comoving_distance_z1z2(
-                z_ls[i], z_next, params
-            )
-            D_is = self.cosmology.transverse_comoving_distance_z1z2(
-                z_ls[i], z_s, params
-            )
-            D_next = self.cosmology.transverse_comoving_distance(z_next, params)
-            alpha_x, alpha_y = self.lenses[i].physical_deflection_angle(
-                X * rad_to_arcsec / D_l,
-                Y * rad_to_arcsec / D_l,
-                z_s,
-                params,
+        for plane in planes:
+            z_l = z_ls[plane]
+            z_next = z_ls[plane + 1] if plane != len(planes) - 1 else z_s
+            alpha_x, alpha_y = self.lenses[plane].reduced_deflection_angle(
+                theta_x, theta_y, z_next, params
             )
 
-            # Update angle of rays after passing through lens (sum in eq 18)
+            # Time delay
+            if any(
+                [
+                    return_shapiro_time_delay,
+                    return_geometric_time_delay,
+                    return_time_delay,
+                ]
+            ):
+                D_l = self.cosmology.angular_diameter_distance(z_l)
+                D_next = self.cosmology.angular_diameter_distance(z_next)
+                D_l_next = self.cosmology.angular_diameter_distance_z1z2(z_l, z_next)
[…]
