+++ src/caustics/sims/lens_source.py
@@ -22,26 +22,23 @@ __all__ = ("LensSource",)
 class LensSource(Simulator):
     """Lens image of a source.
 
-    Straightforward simulator to sample a lensed image of a source
-    object. Constructs a sampling grid internally based on the
-    pixelscale and gridding parameters. It can automatically upscale
-    and fine sample an image. This is the most straightforward
-    simulator to view the image if you already have a lens and source
-    chosen.
+    Straightforward simulator to sample a lensed image of a source object.
+    Constructs a sampling grid internally based on the pixelscale and gridding
+    parameters. It can automatically upscale and fine sample an image. This is
+    the most straightforward simulator to view the image if you already have a
+    lens and source chosen.
 
     Example usage::
 
-       import matplotlib.pyplot as plt
-       import caustics
+       import matplotlib.pyplot as plt import caustics
 
-       cosmo = caustics.FlatLambdaCDM()
-       lens = caustics.lenses.SIS(cosmology = cosmo, x0 = 0., y0 = 0., th_ein = 1.)
-       source = caustics.sources.Sersic(x0 = 0., y0 = 0., q = 0.5, phi = 0.4, n = 2., Re = 1., Ie = 1.)
-       sim = caustics.sims.LensSource(lens, source, pixelscale = 0.05, gridx = 100, gridy = 100, upsample_factor = 2, z_s = 1.)
+       cosmo = caustics.FlatLambdaCDM() lens = caustics.lenses.SIS(cosmology =
+       cosmo, x0 = 0., y0 = 0., th_ein = 1.) source = caustics.sources.Sersic(x0
+       = 0., y0 = 0., q = 0.5, phi = 0.4, n = 2., Re = 1., Ie = 1.) sim =
+       caustics.sims.LensSource(lens, source, pixelscale = 0.05, gridx = 100,
+       gridy = 100, upsample_factor = 2, z_s = 1.)
 
-       img = sim()
-       plt.imshow(img, origin = "lower")
-       plt.show()
+       img = sim() plt.imshow(img, origin = "lower") plt.show()
 
     Attributes
     ----------
@@ -56,13 +53,27 @@ class LensSource(Simulator):
     lens_light: Source, optional
         caustics light object which defines the lensing object's light
     psf: Tensor, optional
-        An image to convolve with the scene. Note that if ``upsample_factor > 1`` the psf must also be at the higher resolution.
+        An image to convolve with the scene. Note that if ``upsample_factor >
+        1`` the psf must also be at the higher resolution.
     pixels_y: Optional[int]
-        number of pixels on the y-axis for the sampling grid. If left as ``None`` then this will simply be equal to ``gridx``
+        number of pixels on the y-axis for the sampling grid. If left as
+        ``None`` then this will simply be equal to ``gridx``
     upsample_factor (default 1)
-        Amount of upsampling to model the image. For example ``upsample_factor = 2`` indicates that the image will be sampled at double the resolution then summed back to the original resolution (given by pixelscale and gridx/y).
+        Amount of upsampling to model the image. For example ``upsample_factor =
+        2`` indicates that the image will be sampled at double the resolution
+        then summed back to the original resolution (given by pixelscale and
+        gridx/y).
+    quad_level: int (default None)
+        sub pixel integration resolution. This will use Gaussian quadrature to
+        sample the image at a higher resolution, then integrate the image back
+        to the original resolution. This is useful for high accuracy integration
+        of the image, but may increase memory usage and runtime.
     psf_pad: Boolean(default True)
-        If convolving the PSF it is important to sample the model in a larger FOV equal to half the PSF size in order to account for light that scatters from outside the requested FOV inwards. Internally this padding will be added before sampling, then cropped off before returning the final image to the user.
[…]
