### 2024-07-26 63c21a5
update unit tests for LensSource
