### 2024-07-26 a878514
refactor: lens source gridding system
