### 2024-07-26 b110c8c
fix no source mu shape
