### 2024-07-26 8d742f9
fix example usage docstring
