+++ src/caustics/sims/lens_source.py
@@ -74,12 +74,6 @@ class LensSource(Simulator):
         sample the image at a higher resolution, then integrate the image back
         to the original resolution. This is useful for high accuracy integration
         of the image, but may increase memory usage and runtime.
-    psf_pad: Boolean(default True)
-        If convolving the PSF it is important to sample the model in a larger
-        FOV equal to half the PSF size in order to account for light that
-        scatters from outside the requested FOV inwards. Internally this padding
-        will be added before sampling, then cropped off before returning the
-        final image to the user.
     z_s: optional
         redshift of the source
     name: string (default "sim")
@@ -126,19 +120,21 @@ class LensSource(Simulator):
             Optional[Source],
             "caustics light object which defines the lensing object's light",
         ] = None,
-        psf: Annotated[Optional[Tensor], "An image to convolve with the scene"] = None,
         pixels_y: Annotated[
             Optional[int], "number of pixels on the y-axis for the sampling grid"
         ] = None,
         upsample_factor: Annotated[int, "Amount of upsampling to model the image"] = 1,
         quad_level: Annotated[Optional[int], "sub pixel integration resolution"] = None,
-        psf_pad: Annotated[bool, "Flag to apply padding to psf"] = True,
         psf_mode: Annotated[
-            Literal["fft", "conv2d"], "Mode for convolving psf"
+            Literal["off", "fft", "conv2d"], "Mode for convolving psf"
         ] = "fft",
+        psf_shape: Annotated[Optional[tuple[int, ...]], "The shape of the psf"] = None,
         z_s: Annotated[
             Optional[Union[Tensor, float]], "Redshift of the source", True
         ] = None,
+        psf: Annotated[
+            Optional[Union[Tensor, list]], "An image to convolve with the scene", True
+        ] = [[1.0]],
         x0: Annotated[
             Optional[Union[Tensor, float]],
             "center of the fov for the lens source image",
@@ -157,12 +153,19 @@ class LensSource(Simulator):
         self.lens = lens
         self.source = source
         self.lens_light = lens_light
-        if psf is None:
-            self.psf = None
+
+        # Configure PSF
+        if psf == [[1.0]]:
+            self._psf_mode = "off"
         else:
-            self.psf = torch.as_tensor(psf)
-            self.psf /= psf.sum()  # ensure normalized
+            self._psf_mode = psf_mode
+        if psf is not None:
+            psf = torch.as_tensor(psf)
+        self._psf_shape = psf.shape if psf is not None else psf_shape
+
+        # Build parameters
         self.add_param("z_s", z_s)
+        self.add_param("psf", psf, self.psf_shape)
         self.add_param("x0", x0)
         self.add_param("y0", y0)
         self._pixelscale = pixelscale
@@ -170,30 +173,18 @@ class LensSource(Simulator):
         # Image grid
         self._pixels_x = pixels_x
         self._pixels_y = pixels_x if pixels_y is None else pixels_y
-
-        # PSF padding if needed
-        self.psf_mode = psf_mode
-        if psf_pad and self.psf is not None:
-            self.psf_pad = (self.psf.shape[1] // 2 + 1, self.psf.shape[0] // 2 + 1)
-        else:
-            self.psf_pad = (0, 0)
-
-        # Build the imaging grid
         self._upsample_factor = upsample_factor
         self._quad_level = quad_level
-        self._build_grid()
 
-        if self.psf is not None:
-            self.psf_fft = self._fft2_padded(self.psf)
[…]
