### 2024-07-26 ea6d6d0
get outputs right, shape and normalization
