### 2024-07-26 1c9e9c6
add check conv2d vs fft
