### 2024-07-26 f2e1c0e
remove unnecessary 'off' option for psf_mode
