### 2024-07-30 #6922 @ivastar
@ConnorStoneAstro I am taking over as editor. I will be looking for reviewers in the next week or so. Please let me know if you have suggestions for appropriate reviewers, here or via e-mail.
