### 2024-07-30 #6922 @ConnorStoneAstro
Hi @ivastar that's great! There are many people who could give excellent reviews including: Simon Birrer, Andi Gu, Robert Morgan, James Nightingale, R. Benton Metcalf, Margarita Petkova, and Sebastian Wagner-Carena. They have all been involved in gravitational lensing software projects, and are cited in the paper.
