### 2024-08-05 #6922 @ivastar
@rmorgan10, @swagnercarena would you be interested in reviewing this submission for JOSS?

The submission repository is at:  [https://github.com/Ciela-Institute/caustics](https://github.com/Ciela-Institute/caustics) You can see the full submission here: https://github.com/openjournals/joss-reviews/issues/6922 (this issue)

JOSS is a free, open-source, community driven and developer-friendly online journal.

The review process at JOSS place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. JOSS reviews involve downloading and installing the software, and inspecting the repository and submitted paper for key elements. See [Review Criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)

Editors and reviewers post comments on the Review issue, and authors respond to the comments and improve their submission until acceptance (or withdrawal, if they feel unable to satisfy the review). We aim for reviews to be completed within about 3 weeks (~ end of August for this submission).
