### 2024-08-07 #7081 @editorialbot
[REVIEW]: Caustics: A Python Package for Accelerated Strong Gravitational Lensing Simulations
**Submitting author:** @ConnorStoneAstro (<a href="http://orcid.org/0000-0002-9086-6398">Connor Stone</a>)
**Repository:** https://github.com/Ciela-Institute/caustics
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
