### 2024-08-07 9d568e8
Fix is None pre-commit check
