### 2024-08-13 dd478d2
getting pixelated convergence to remove artifacts
