### 2024-08-13 d41e10f
no more artifacts in pixelated convergence
