### 2024-08-15 9cf12a9
remove half pixel shift in kernel
