### 2024-08-17 e0e8070
Change kernel window to be round
