### 2024-08-28 dcd15f8
return fixes to pixelated convergence
