+++ docs/source/_toc.yml
@@ -26,6 +26,7 @@ chapters:
     sections:
       - file: examples/Example_ImageFit_LM
       - file: examples/Example_ImageFit_NUTS
+      - file: examples/Example_ImageFit_EMCEE
       - file: examples/Example_QSOLensFit
   - file: contributing
   - file: frequently_asked_questions
