### 2024-08-29 d906415
footnote formatting
