### 2024-08-30 346d0ef
update comments in mcmc example
