### 2024-08-31 902d1ef
Add MH and MALA MCMC examples (#251)
* Adding emcee example

* add discussion

* Add notebook to toc

* reduce emcee samples

* merge emcee example into NUTS example

* update comments in mcmc example

* change cell tags

* add tags to all tutorials

* add mala example to MCMC notebook

* filename change

* collapse chains into overlap with different colors

* fix filenames
