+++ .github/workflows/cd.yml
@@ -57,7 +57,7 @@ jobs:
           ls -ltrh
           ls -ltrh dist
       - name: Publish to Test PyPI
-        uses: pypa/gh-action-pypi-publish@v1.9.0
+        uses: pypa/gh-action-pypi-publish@v1.10.0
         with:
           repository-url: https://test.pypi.org/legacy/
           verbose: true
@@ -95,5 +95,5 @@ jobs:
           name: artifact
           path: dist
 
-      - uses: pypa/gh-action-pypi-publish@v1.9.0
+      - uses: pypa/gh-action-pypi-publish@v1.10.0
         if: startsWith(github.ref, 'refs/tags')
