+++ .pre-commit-config.yaml
@@ -51,7 +51,7 @@ repos:
         args: [--prose-wrap=always]
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: "v0.6.2"
+    rev: "v0.6.3"
     hooks:
       - id: ruff
         args: ["--fix", "--show-fixes"]
