### 2024-09-03 48ea3ae
docs: Put logo gif in main docs page (#243)
* Put logo gif in main docs page

* silly typo fix
