### 2024-09-06 97561c5
use pytest mark parametrize to massively increase the number of tests
