+++ .github/workflows/ci.yml
@@ -74,8 +74,9 @@ jobs:
         if:
           ${{ matrix.python-version == '3.10' && matrix.os == 'ubuntu-latest'}}
         uses: codecov/codecov-action@v4
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         with:
-          token: ${{ secrets.CODECOV_TOKEN }}
           files: ./coverage.xml
           flags: unittests
           name: codecov-umbrella
