### 2024-09-12 e154ff8
use pytest mark parametrize to massively increase the number of tests (#255)
