+++ tests/test_nfw.py
@@ -14,12 +14,16 @@ from utils import lens_test_helper, setup_grids
 from caustics.cosmology import FlatLambdaCDM as CausticFlatLambdaCDM
 from caustics.lenses import NFW
 
+import pytest
+
 h0_default = float(default_cosmology.get().h)
 Om0_default = float(default_cosmology.get().Om0)
 Ob0_default = float(default_cosmology.get().Ob0)
 
 
-def test(sim_source, device, lens_models):
+@pytest.mark.parametrize("m", [1e8, 1e10, 1e12])
+@pytest.mark.parametrize("c", [1.0, 8.0, 20.0])
+def test_nfw(sim_source, device, lens_models, m, c):
     atol = 1e-5
     rtol = 3e-2
     z_l = torch.tensor(0.1)
@@ -54,8 +58,8 @@ def test(sim_source, device, lens_models):
 
     thx0 = 0.457
     thy0 = 0.141
-    m = 1e12
-    c = 8.0
+    # m = 1e12
+    # c = 8.0
     x = torch.tensor([thx0, thy0, m, c])
 
     # Lenstronomy
@@ -113,7 +117,3 @@ def test_runs(sim_source, device, lens_models):
     assert torch.all(torch.isfinite(alpha[1]))
     kappa = lens.convergence(thx, thy, z_s, x)
     assert torch.all(torch.isfinite(kappa))
-
-
-if __name__ == "__main__":
-    test()
