+++ src/caustics/light/func/__init__.py
@@ -1,3 +1,4 @@
 from .sersic import brightness_sersic, k_lenstronomy, k_sersic
+from .point_source import brightness_point
 
 __all__ = ("brightness_sersic", "k_lenstronomy", "k_sersic")
