### 2024-09-12 0f74749
Add both images and magnification
