### 2024-09-16 ba1d555
Add limb darkening and merge image+mag
