+++ src/caustics/light/func/point_source.py
@@ -4,11 +4,22 @@ from ...utils import translate_rotate
 
 
 
-def brightness_point(x0, y0, theta_s, Ie, x, y, s=0.0):
+def brightness_point(x0, y0, theta_s, Ie, x, y, gamma=1.0, s=0.0):
 
     x, y = translate_rotate(x, y, x0, y0, phi=None) # phi=None because a point source is spherically symmetric
+
+    # todo add linear limb darkening 
+    # Calculate the radial distance from the center in units of the stellar radius
+    impact_parameter = (x**2 + y**2)**(1/2) / theta_s
+    #linear limb darkening
+    mu = (1 - impact_parameter**2)**(1/2)
+    intensity = Ie * (1 - gamma * (1 - mu))
+    return intensity
     
-    mask = (x**2 + y**2)**(1/2) <= theta_s #Uniform surface brightness inside source
+    # mask = (x**2 + y**2)**(1/2) <= theta_s #Uniform surface brightness inside source
     
-    # Return Ie where the mask is True, otherwise return 0
-    return Ie * mask.float()
+    # # Return Ie where the mask is True, otherwise return 0
+    # return Ie * mask.float()
+
+
+#
