### 2024-09-16 2ce7418
Adding PointSource to __all__
