+++ .pre-commit-config.yaml
@@ -69,7 +69,13 @@ repos:
     rev: "v2.3.0"
     hooks:
       - id: codespell
-        args: ["--write-changes", "--ignore-words", ".codespell-whitelist"]
+        args:
+          [
+            "--write-changes",
+            "--ignore-words",
+            ".codespell-whitelist",
+            "--ignore-words-list=tE",
+          ]
 
   #- repo: https://github.com/kynan/nbstripout
   #rev: 0.7.1
