### 2024-09-24 Ciela-Institute/caustics#262 @Jammy2211
JOSS Review: Point Source Docs + Expand on limitations
It would be good if the following page had a bit more information on how the lens equation is inverted:

https://caustics.readthedocs.io/en/latest/tutorials/InvertLensEquation.html

**Methodology:**

I had to go to the source code to learn it is done by throwing down points in the image-plane, and then running an optimizer (correct me if I'm wrong).

This is quite different to how many lensing codes work, which instead iteratively ray trace triangles to the image plane and keep those triangles which contain the source centre. I suspect many users from a lensing background would expect this to be the default option.

A visualization showing the process would help a lot (e.g. an image with the initial points, then an image with the final points).

**Limitations**

All lens equation solvers have limitations, and documenting these clearly is important to ensuring users do not misuse them for science.

For example, in the example above, the central fifth image of the SIE mass profile is not located nor is it mentioned. Whilst it is heavily demagnified (and therefore any emission from an extended source would likely not be observed), the image does still technically exist and I would expect the default behavior to be that the lens equation solver finds it (or has the user provide an input which explicitly removes it).

I think it would be good to add a few of the more extreme cases of lensing which typically break lens inverters, so that a user can understand where the code might infer an incorrect solution. For example, when the source is on the edge of a caustic that two of two multiple images in the image-plane are very close to one another, and thus the solver has to successfully "split" them at sub-pixel resolution. I am guessing whether your solver does this or not successfully depends on the random draw of the image-plane coordinates? Or is something more clever going on?
