### 2024-09-24 Ciela-Institute/caustics#263 @Jammy2211
JOSS Review: Mention Herculens
Can the JOSS paper be updated to specifically refer to Herculens (https://github.com/Herculens/herculens?tab=readme-ov-file) in the the second paragraph of the "statement of need" section.

This software uses JAX and is probably the closest open source package to `caustics`, albeit I don't see dedication documentation for lensing simulations.
