### 2024-09-24 #7081 @Jammy2211
Review complete, put two minor issues up which once complete mean I can recommend for acceptance:

https://github.com/Ciela-Institute/caustics/issues/263
https://github.com/Ciela-Institute/caustics/issues/262
