+++ src/caustics/light/func/star_source.py
@@ -3,13 +3,13 @@ from torch import where
 from ...utils import translate_rotate
 
 
-def brightness_point(x0, y0, theta_s, Ie, x, y, gamma=1.0, s=0.0):
+def brightness_star(x0, y0, theta_s, Ie, x, y, gamma=0.0):
 
     x, y = translate_rotate(
-        x, y, x0, y0, phi=None
-    )  # phi=None because a point source is spherically symmetric
+        x, y, x0, y0
+    ) 
 
-    # Calculate the radial distance from the center in units of the stellar radius
+    # Calculate the radial distance from the center
     impact_parameter = (x**2 + y**2) ** (1 / 2)
     # linear limb darkening
     mu = (1 - impact_parameter**2) ** (1 / 2)
