+++ src/caustics/light/func/__init__.py
@@ -1,4 +1,4 @@
 from .sersic import brightness_sersic, k_lenstronomy, k_sersic
-from .point_source import brightness_point
+from .star_source import brightness_star
 
-__all__ = ("brightness_sersic", "k_lenstronomy", "k_sersic", "brightness_point")
+__all__ = ("brightness_sersic", "k_lenstronomy", "k_sersic", "brightness_star")
