+++ src/caustics/light/__init__.py
@@ -2,6 +2,6 @@ from .base import Source
 from .pixelated import Pixelated
 from .sersic import Sersic
 from .pixelated_time import PixelatedTime
-from .point_source import PointSource
+from .star_source import StarSource
 
-__all__ = ["Source", "Pixelated", "PixelatedTime", "Sersic", "PointSource"]
+__all__ = ["Source", "Pixelated", "PixelatedTime", "Sersic", "StarSource"]
