### 2024-09-25 Ciela-Institute/caustics#262 @ConnorStoneAstro
@Jammy2211 Thanks for catching this! We are going to look into the forward raytracing, make sure it catches all the images, and expand on the documentation. I'll comment here when its ready!
