### 2024-09-25 Ciela-Institute/caustics#263 @ConnorStoneAstro
Thanks! I'll make sure it's in the updated version
