### 2024-09-25 fac7eb7
starting to make recursive triangle forward raytrace
