### 2024-09-26 7a43dc4
Update Microlensing/simulator tutorial and toc. Remove fitting
