+++ docs/source/_toc.yml
@@ -22,6 +22,7 @@ chapters:
       - file: tutorials/Parameters
       - file: tutorials/Simulators
       - file: tutorials/Playground
+      - file: tutorials/Microlensing
   - file: examples/index
     sections:
       - file: examples/Example_ImageFit_LM
