### 2024-09-26 6b5dbcf
fix invert lens demo to go all the way to convergence.
