### 2024-09-26 ecbda4e
Improve forward raytrace to better find all images (#264)
* starting to make recursive triangle forward raytrace

* updated tutorial for inverting the lens equation

* hide plotting cell

* set sampling to uniform for naive method

* remove unconverged points in naive example

* fix invert lens demo to go all the way to convergence.

* fix demagification value
