+++ src/caustics/lenses/func/__init__.py
@@ -1,5 +1,12 @@
 from .base import (
     forward_raytrace,
+    triangle_contains,
+    triangle_area,
+    triangle_neighbors,
+    triangle_upsample,
+    triangle_equals,
+    remove_triangle_duplicates,
+    forward_raytrace_rootfind,
     physical_from_reduced_deflection_angle,
     reduced_from_physical_deflection_angle,
     time_delay_arcsec2_to_days,
@@ -60,6 +67,13 @@ from .enclosed_mass import (
 
 __all__ = (
     "forward_raytrace",
+    "triangle_contains",
+    "triangle_area",
+    "triangle_neighbors",
+    "triangle_upsample",
+    "triangle_equals",
+    "remove_triangle_duplicates",
+    "forward_raytrace_rootfind",
     "physical_from_reduced_deflection_angle",
     "reduced_from_physical_deflection_angle",
     "time_delay_arcsec2_to_days",
