+++ src/caustics/lenses/func/base.py
@@ -4,37 +4,135 @@ from ...utils import batch_lm
 from ...constants import arcsec_to_rad, c_Mpc_s, days_to_seconds
 
 
-def forward_raytrace(bx, by, raytrace, epsilon, n_init, fov):
+def triangle_contains(p, v):
     """
-    Perform a forward ray-tracing operation which maps from the source plane to the image plane.
+    determine if point v is inside triangle p. Where p is a (3,2) tensor, and v
+    is a (2,) tensor.
+    """
+    p01 = p[1] - p[0]
+    p02 = p[2] - p[0]
+    dp0p02 = p[0][0] * p02[1] - p[0][1] * p02[0]
+    dp0p01 = p[0][0] * p01[1] - p[0][1] * p01[0]
+    dp01p02 = p01[0] * p02[1] - p01[1] * p02[0]
+    dvp02 = v[0] * p02[1] - v[1] * p02[0]
+    dvp01 = v[0] * p01[1] - v[1] * p01[0]
+    a = (dvp02 - dp0p02) / dp01p02
+    b = -(dvp01 - dp0p01) / dp01p02
+    return (a >= 0) & (b >= 0) & (a + b <= 1)
+
+
+def triangle_area(p):
+    """
+    Determine the area of triangle p where p is a (3,2) tensor.
+    """
+    return (
+        0.5
+        * (
+            p[0][0] * (p[1][1] - p[2][1])
+            + p[1][0] * (p[2][1] - p[0][1])
+            + p[2][0] * (p[0][1] - p[1][1])
+        ).abs()
+    )
+
+
+def triangle_neighbors(p):
+    """
+    Build a set of neighbors for triangle p where p is a (3,2) tensor. The
+    neighbors all have the same shape as p, but are various translations and
+    reflections of p that share a common edge or vertex.
+    """
+    p01 = p[1] - p[0]
+    p02 = p[2] - p[0]
+    p12 = p[2] - p[1]
+    pref = -(p - p[0]) + p[0]
+    return torch.stack(
+        (
+            p,
+            p + p01,
+            p - p01,
+            p + p02,
+            p - p02,
+            p + p12,
+            p - p12,
+            pref,
+            pref + p01,
+            pref + 2 * p01,
+            pref + p02,
+            pref + 2 * p02,
+            pref + p01 + p02,
+        ),
+        dim=0,
+    )
+
+
+def triangle_upsample(p):
+    """
+    Upsample triangle p where p is a (3,2) tensor. The upsampled triangles are
+    all triangles internal to p built by taking the midpoints of the edges of p.
+    """
+    p01 = (p[1] + p[0]) / 2
+    p02 = (p[2] + p[0]) / 2
+    p12 = (p[2] + p[1]) / 2
+    return torch.stack(
+        (
+            torch.stack((p[0], p01, p02), dim=0),
+            torch.stack((p01, p[1], p12), dim=0),
+            torch.stack((p02, p12, p[2]), dim=0),
+            torch.stack((p01, p12, p02), dim=0),
+        ),
+        dim=0,
+    )
+
+
+def triangle_equals(p1, p2):
+    """
+    Determine if two triangles are equal. Where p1 and p2 are (3,2) tensors.
+    """
+    return torch.all((p1 - p2).abs() < 1e-6)
+
+
+def remove_triangle_duplicates(p):
+    unique_triangles = torch.zeros((0, 3, 2))
+    B = p.shape[0]
+    batch_triangle_equals = torch.vmap(triangle_equals, in_dims=(None, 0))
+    for i in range(B):
+        # Compare current triangle with all triangles in the unique list
+        if i == 0 or not batch_triangle_equals(p[i], unique_triangles).any():
[…]
