+++ paper.bib
@@ -24,6 +24,24 @@ title = {`PyAutoLens`: Open-Source Strong Gravitational Lensing},
 journal = {Journal of Open Source Software}
 }
 
+@ARTICLE{Herculens,
+       author = {{Galan}, A. and {Vernardos}, G. and {Peel}, A. and {Courbin}, F. and {Starck}, J. -L.},
+        title = "{Using wavelets to capture deviations from smoothness in galaxy-scale strong lenses}",
+      journal = {\aap},
+     keywords = {galaxies: structure, dark matter, methods: data analysis, gravitation, gravitational lensing: strong, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+         year = 2022,
+        month = dec,
+       volume = {668},
+          eid = {A155},
+        pages = {A155},
+          doi = {10.1051/0004-6361/202244464},
+archivePrefix = {arXiv},
+       eprint = {2207.05763},
+ primaryClass = {astro-ph.IM},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022A&A...668A.155G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{GIGALens,
        author = {{Gu}, A. and {Huang}, X. and {Sheu}, W. and {Aldering}, G. and {Bolton}, A.~S. and {Boone}, K. and {Dey}, A. and {Filipp}, A. and {Jullo}, E. and {Perlmutter}, S. and {Rubin}, D. and {Schlafly}, E.~F. and {Schlegel}, D.~J. and {Shu}, Y. and {Suyu}, S.~H.},
         title = "{GIGA-Lens: Fast Bayesian Inference for Strong Gravitational Lens Modeling}",
@@ -330,7 +348,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Millon2024,
-       author = {{Millon}, Martin and {Michalewicz}, Kevin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric and {Marshall}, Philip J.},
+       author = {{Million}, Martin and {Michalewicz}, Kevin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric and {Marshall}, Philip J.},
         title = "{Image deconvolution and PSF reconstruction with STARRED: a wavelet-based two-channel method optimized for light curve extraction}",
       journal = {arXiv e-prints},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
@@ -449,7 +467,7 @@ url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-perf
 }
 
 @software{JAX,
-  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
+  author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leery and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
   url = {http://github.com/google/jax},
   version = {0.3.13},
