### 2024-09-26 756ac24
minor wording updates
