+++ paper.md
@@ -91,8 +91,9 @@ measurement of the Universe's expansion rate [e.g. @holycow], dark matter [e.g.
 future surveys expected to discover hundreds of thousands of lensing systems,
 the modelling and simulation of such systems must occur at orders of magnitude
 larger scale then ever before. Here we present `caustics`, a Python package
-designed to handle the extensive computational demands of modeling such a vast
-number of lensing systems.
+designed to facilitate machine learning and Bayesian methods to handle the
+extensive computational demands of modeling such a vast number of lensing
+systems.
 
 # Statement of need
 
@@ -132,7 +133,7 @@ expertise but allows more capabilities. In this way, users with all levels of
 gravitational lensing simulation experience may effectively engage with the
 software. Flexibility is achieved by a determined focus on minimalism in the
 core functionality of `caustics`. All of these elements combine to make
-`Caustics` a capable lensing simulator to support machine learning applications,
+`caustics` a capable lensing simulator to support machine learning applications,
 and classical analysis.
 
 `Caustics` fulfills a timely need for a differentiable lensing simulator.
@@ -149,14 +150,17 @@ over 100,000 lenses in a timely manner [@Hezaveh2017; @Perreault2017].
 
 `Caustics` is a gravitational lensing simulator. The purpose of the project is
 to streamline the simulation of strong gravitational lensing effects on the
-light of a background source. This includes a variety of parametric lensing
-profiles such as: Singular Isothermal Ellipsoid (SIE), Elliptical Power Law
-(EPL), Pseudo-Jaffe, Navarro-Frenk-White (NFW), and External Shear.
-Additionally, it offers non-parametric representations such as a gridded
-convergence or potential field. For the background source `caustics` provides a
-Sérsic light profile, as well as a pixelized light image. Users may easily
-extend these lens and source lists using templates provided in the
-documentation.
+light of a background source. Primarily the focus is on all transformations
+between the source plane(s) and the image plane through the lensing plane(s).
+There is minimal effort on modelling the observational elements of atmosphere,
+telescope optics, detector effects with most of these left to the user. A
+variety of parametric lensing profiles are included, such as: Singular
+Isothermal Ellipsoid (SIE), Elliptical Power Law (EPL), Pseudo-Jaffe,
+Navarro-Frenk-White (NFW), and External Shear. Additionally, it offers
+non-parametric representations such as a gridded convergence or potential field.
+For the background source `caustics` provides a Sérsic light profile, as well as
+a pixelized light image. Users may easily extend these lens and source lists
+using templates provided in the documentation.
 
 Once a lensing system has been defined, `caustics` may then perform various
 computational operations on the system such as raytracing through the lensing
@@ -198,7 +202,8 @@ implemented, though they represent an avenue for future development.
 functionality to optimize or sample the resulting functions. Users are
 encouraged to use already existing optimization and sampling codes like
 `scipy.optimize` [@scipy], `emcee` [@emcee], `dynesty` [@dynesty], `Pyro`
-[@pyro], and `torch.optim` [@pytorch].
+[@pyro], and `torch.optim` [@pytorch]. Interfacing with these codes is easy and
+demonstrations are included in the documentation.
 
 # Performance
 
@@ -261,10 +266,10 @@ computing FFTs for convolution, plus PyTorch overhead, this will fill the GPU.
 An A100 GPU with 80GB of memory would be able to go much further before
 saturating, staying in the flat scaling region longer and giving even further
 performance improvements over CPU computations. Nonetheless, it is possible to
-easily achieve over 100X speedup over CPU performance, making GPUs by far the
-most efficient method to perform large lensing computations such as running many
-MCMC chains or sampling many lensing realizations (e.g. for training machine
[…]
