### 2024-09-30 Ciela-Institute/caustics#263 @ConnorStoneAstro
Indeed, Herculens is an excellent addition to the literature review at the beginning of the paper. We have added a citation to the statement of need section alongside the other lensing codes. You can see the updated draft in the [JOSS review page](https://github.com/openjournals/joss-reviews/issues/7081). Please let me know if there is anything else you think we should comment on!
