### 2024-09-30 Ciela-Institute/caustics#262 @ConnorStoneAstro
Thanks again for spotting this, it was fun to code up! We have updated the forward_raytrace function to use the triangle method you mentioned, with a root finder to finalize the position. Please see the updated docs on inverting the lens equation and let me know if there is anything else you think we should add!

https://caustics.readthedocs.io/en/latest/tutorials/InvertLensEquation.html
