### 2024-09-30 082cba5
adding unit tests
