+++ docs/source/_toc.yml
@@ -16,12 +16,12 @@ chapters:
           - file: tutorials/InterfaceIntroduction_oop
           - file: tutorials/InterfaceIntroduction_func
       - file: tutorials/LensZoo
+      - file: tutorials/Microlensing
       - file: tutorials/VisualizeCaustics
       - file: tutorials/MultiplaneDemo
       - file: tutorials/InvertLensEquation
       - file: tutorials/Parameters
       - file: tutorials/Simulators
-      - file: tutorials/Microlensing
   - file: examples/index
     sections:
       - file: examples/Example_ImageFit_LM
