### 2024-10-02 da8d8eb
remove excess files, clean up notebook
