### 2024-10-02 7ce5d59
Microlensing (#259)
* init

* Minimal microlensing example

* Adding point source

* Add both images and magnification

* Fun animation

* fitting working

* mcmc working

* Add limb darkening and merge image+mag

* added limb darkening

* Add limb darkening functionality

* Improving microlensing fit example

* fixing pre-commit

* Adding PointSource to __all__

* adding tE to codespell ignore

* Fix bug and improve microlensing fit example

* style: pre-commit fixes

* Noting issue with PyLIMA

* strip nbs

* Fixed bug in u0 units

* Adjusting mcmc

* changed pointsource to starsource, fixin
[…]
