### 2024-10-10 9f293c8
Added light stack module
