+++ src/caustics/light/__init__.py
@@ -1,7 +1,8 @@
 from .base import Source
 from .pixelated import Pixelated
 from .sersic import Sersic
+from .light_stack import LightStack
 from .pixelated_time import PixelatedTime
 from .star_source import StarSource
 
-__all__ = ["Source", "Pixelated", "PixelatedTime", "Sersic", "StarSource"]
+__all__ = ["Source", "Pixelated", "PixelatedTime", "Sersic", "LightStack", "StarSource"]
