+++ src/caustics/__init__.py
@@ -30,6 +30,7 @@ from .light import (
     Pixelated,
     PixelatedTime,
     Sersic,
+    LightStack,
     StarSource,
 )
 from . import utils
@@ -67,6 +68,7 @@ __all__ = [
     "Pixelated",
     "PixelatedTime",
     "Sersic",
+    "LightStack",
     "StarSource",
     "utils",
     "LensSource",
