### 2024-10-10 eb81bba
separate pixelated kernel and windowing into different functions
