### 2024-10-10 1a72158
allow multiple multipole moments. add unit tests
