+++ tests/test_multipole.py
@@ -1,7 +1,7 @@
 import torch
 import numpy as np
 from lenstronomy.LensModel.lens_model import LensModel
-from .utils import (
+from utils import (
     alpha_test_helper,
     kappa_test_helper,
     Psi_test_helper,
@@ -32,6 +32,7 @@ def test_multipole_lenstronomy(sim_source, device, lens_models, m_order):
                 z_l: {float(z_l)}
             init_kwargs:
                 cosmology: *cosmology
+                m: {m_order}
         """
         yaml_dict = yaml.safe_load(yaml_str.encode("utf-8"))
         mod = lens_models.get("Multipole")
@@ -39,13 +40,13 @@ def test_multipole_lenstronomy(sim_source, device, lens_models, m_order):
     else:
         # Models
         cosmology = FlatLambdaCDM(name="cosmo")
-        lens = Multipole(name="multipole", cosmology=cosmology, z_l=z_l)
+        lens = Multipole(name="multipole", cosmology=cosmology, z_l=z_l, m=m_order)
     lens_model_list = ["MULTIPOLE"]
     lens_ls = LensModel(lens_model_list=lens_model_list)
 
     # Parameters  m, a_m ,phi_m
     z_s = torch.tensor(1.2)
-    x = torch.tensor([-0.342, 0.51, 0.1, 3.14 / 4, m_order])
+    x = torch.tensor([-0.342, 0.51, 0.1, 3.14 / 4])
     kwargs_ls = [
         {
             "center_x": x[0].item(),
