+++ .pre-commit-config.yaml
@@ -9,12 +9,12 @@ ci:
 
 repos:
   - repo: https://github.com/psf/black
-    rev: "24.8.0"
+    rev: "24.10.0"
     hooks:
       - id: black-jupyter
 
   - repo: https://github.com/asottile/blacken-docs
-    rev: "1.18.0"
+    rev: "1.19.0"
     hooks:
       - id: blacken-docs
         additional_dependencies: [black==23.7.0]
