### 2024-10-18 5afeab7
feat: add multipole lens profile (#215)
* add multipole lens profile

* typo correction

* update variable names

* type multipole order m checking

* lenstronomy lensmodellist update

* higher tolerance

* higher tolerance rtol and atol

* try without derotation

* do not do kappa test

* adjust; multipole moment can also be 2

* fix convergence by factor 1/2

* style: pre-commit fixes

* allow multiple multipole moments. add unit tests

* tests now pass

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Connor Stone <connorst
[…]
