### 2024-10-18 fb0b0e9
Added light stack module (#272)
