+++ src/caustics/__init__.py
@@ -35,7 +35,7 @@ from .light import (
     StarSource,
 )
 from . import utils
-from .sims import LensSource, Microlens, Simulator
+from .sims import LensSource, Microlens
 from .tests import test
 from .models.api import build_simulator
 from . import func
@@ -75,7 +75,6 @@ __all__ = [
     "utils",
     "LensSource",
     "Microlens",
-    "Simulator",
     "test",
     "build_simulator",
     "func",
