### 2024-10-18 a30dcfb
clear pydantic, some tests pass
