+++ src/caustics/models/utils.py
@@ -1,285 +1,0 @@
-# mypy: disable-error-code="union-attr, valid-type, has-type, assignment, arg-type, dict-item, return-value, misc"
-import typing
-import inspect
-import torch
-from typing import List, Literal, Dict, Annotated, Union, Any, Tuple
-
-try:
-    from pydantic import Field, create_model, field_validator, ValidationInfo
-except ImportError:
-    raise ImportError(
-        "The `pydantic` package is required to use this feature. "
-        "You can install it using `pip install pydantic==2.7`. This package requires rust. Make sure you have the permissions to install the dependencies.\n "
-        "Otherwise, the maintainer can install the package for you, you can then use `pip install --no-index pydantic`"
-    )
-
-from ..parametrized import Parametrized
-from .base_models import Base, Parameters, InitKwargs
-from .registry import get_kind, _registry
-from ..parametrized import ClassParam
-from ..utils import _import_func_or_class, _eval_expression
-
-PARAMS = "params"
-INIT_KWARGS = "init_kwargs"
-
-
-def _get_kwargs_field_definitions(
-    parametrized_class: Parametrized, dependant_models: Dict[str, Any] = {}
-) -> Dict[str, Dict[str, Any]]:
-    """
-    Get the field definitions for the parameters and init_kwargs of a Parametrized class
-
-    Parameters
-    ----------
-    parametrized_class : Parametrized
-        The Parametrized class to get the field definitions for.
-    dependant_models : Dict[str, Any], optional
-        The dependent models to use, by default {}
-        See: https://docs.pydantic.dev/latest/concepts/unions/#nested-discriminated-unions
-
-    Returns
-    -------
-    dict
-        The resulting field definitions dictionary
-    """
-    cls_signature = inspect.signature(parametrized_class)
-    kwargs_field_definitions: Dict[str, Dict[str, Any]] = {PARAMS: {}, INIT_KWARGS: {}}
-    for k, v in cls_signature.parameters.items():
-        if k != "name":
-            anno = v.annotation
-            dtype = anno.__origin__
-            cls_param = ClassParam(*anno.__metadata__)
-            if cls_param.isParam:
-                kwargs_field_definitions[PARAMS][k] = (
-                    dtype,
-                    Field(default=v.default, description=cls_param.description),
-                )
-            # Below is to handle cases for init kwargs
-            elif k in dependant_models:
-                dependant_model = dependant_models[k]
-                if isinstance(dependant_model, list):
-                    # For the multi lens case
-                    # dependent model is wrapped in a list
-                    dependant_model = dependant_model[0]
-                    kwargs_field_definitions[INIT_KWARGS][k] = (
-                        List[dependant_model],
-                        Field([], description=cls_param.description),
-                    )
-                else:
-                    kwargs_field_definitions[INIT_KWARGS][k] = (
-                        dependant_model,
-                        Field(..., description=cls_param.description),
-                    )
-            elif v.default == inspect._empty:
-                kwargs_field_definitions[INIT_KWARGS][k] = (
-                    dtype,
-                    Field(..., description=cls_param.description),
-                )
-            else:
-                kwargs_field_definitions[INIT_KWARGS][k] = (
-                    dtype,
-                    Field(v.default, description=cls_param.description),
-                )
-    return kwargs_field_definitions
-
-
-def create_pydantic_model(
-    cls: "Parametrized | str", dependant_models: Dict[str, type] = {}
-) -> Base:
-    """
-    Create a pydantic model from a Parametrized class.
-
-    Parameters
-    ----------
-    cls : Parametrized | str
-        The Parametrized class to create the model from.
-    dependant_models : Dict[str, type], optional
-        The dependent models to use, by default {}
[…]
