### 2024-10-21 #7081 @ivastar
@ConnorStoneAstro I have a list of editorial comments, please review and implement. Sorry it is so long, this is a pretty long paper. We will proceed to the publications afterwards:

Line 3: There is already a * for the authors that contributed equally, can you change the symbol for the work done at UdeM, etc?
Line 16: in a variety --> at a variety
Line 17: large-scale, hyphenated
Line 18: Universe, capitalized
Line 19: "The lens objects must align on the sky of order" --> "The lens and the lensed object must be aligned on the sky within"
Line 20: 10's --> tens, arcseonds --> arcseconds (typo)
Line 22+: The sentence starting with "Notably", the items in the list are not all the same, maybe something like: "Notably, they are a tool for understanding a wide range of astrophysical phenomena including supernovae, quasars, the first stars and the expansion rate of the Universe."
Line 27: "must occur" --> must be done
Line 28: then --> than
Line 30: modeling --> modelling
Line 32: LSST stands for Large Synoptic Survey Telescope, correct Sky --> Survey and remove survey at the end of the line
Line 33: and THE Euclid wide survey
Line 33: also mention Roman?
Line 41: low level --> low-level
Line 44: caustics will provide --> caustics provides
Line 47: well-established, should be hypenated
Line 48: C++ based --> C++-based
Line 52: can you please add the names of the libraries cited in these references?
Line 53: "in house" --> in-house
Line 54: Caustics development --> The development of caustics
Line 64: fulfills --> fills
Line 65: benefitted --> benefited
Line 67: light curves --> time series analysis
Line 76: Primary --> The primary or The main
Line 82: "a" potential field
Paragraph starting at line 85: replace all occurrences of "may" with "can", also elsewhere throughout the paper, can means the code is capable
Line 86: remove "either"
Line 104: Since --> Because
Line 110: though --> although, also line 184
Line 112: "Caustics defined scope" --> "The scope of caustics"
Line 114: like --> such as
Line 117: Move "Via PyTorch" to the end of the sentence
Line 122: use Intel --> use an Intel
Line 123: use a NVIDIA --> use an NVIDIA
Line 124: show performance --> show the performance; remove "simply"
First paragraph of the Performance section: Please include explanations of the batched, unbatched and 4 CPU tests.
Lines 131, 134, 151 and Figure 2: CPU should be in all caps
Also Figure 2: can you please move the legend to the upper left corner?
Line 130: lensastronomy should be lowercase
Line 131: Python should be capitalized, also line 143 and please check the rest of the text
Line 146, 147: enhancements --> improvements
Line 147: overhead --> overheads
Line 152: allows --> allow
Paragraph starting at line 150: verbs alternate between past and present tense, go all present
Section on user experience: Sentences alternate between "users" (plural) and "the user" singular, pick one and make it consistent (also the corresponding verbs), I recommend the plural
Line 154: comma after scenario
Line 159: storage --> memory
Line 162: saturating --> saturation
Line 178 and throughout: Object oriented --> Object-oriented
Line 183: with these --> to these
Line 186: ".yaml" --> "configuration file", keeps the terminology consistent
Line 189: add a comma after default
Line 191: "eschew" --> "avoid"
Foonote 1: Capitalize "Kernalizing", "growth" --> "development"
Line 201: since --> because, observe --> see
Line 216: it is simple to --> it is easy to
Line 220: long lasting --> long-lasting
Line 221: poised --> positioned
Line 243: users transition --> the transition of users
Line 254: The work of A.A. and R.L. were --> The work of A.A. and R.L. was

Finally, in the bibliography we don't support short journal names line \apj. Can you please write out the full journal names?
