+++ tests/conftest.py
@@ -2,7 +2,6 @@ import sys
 import os
 import torch
 import pytest
-import typing
 
 # Add the helpers directory to the path so we can import the helpers
 sys.path.append(os.path.join(os.path.dirname(__file__), "utils"))
@@ -10,28 +9,6 @@ sys.path.append(os.path.join(os.path.dirname(__file__), "utils"))
 # from caustics.models.utils import setup_pydantic_models
 
 CUDA_AVAILABLE = torch.cuda.is_available()
-
-LIGHT_ANNOTATED, LENSES_ANNOTATED = None, None  # setup_pydantic_models()
-
-
-def _get_models(annotated):
-    typehint = typing.get_args(annotated)[0]
-    pydantic_models = typing.get_args(typehint)
-    if isinstance(pydantic_models, tuple):
-        pydantic_models = {m.__name__: m for m in pydantic_models}
-    else:
-        pydantic_models = {pydantic_models.__name__: pydantic_models}
-    return pydantic_models
-
-
-@pytest.fixture
-def light_models():
-    return _get_models(LIGHT_ANNOTATED)
-
-
-@pytest.fixture
-def lens_models():
-    return _get_models(LENSES_ANNOTATED)
 
 
 @pytest.fixture(params=["yaml", "no_yaml"])
