+++ requirements.txt
@@ -1,5 +1,5 @@
 astropy>=5.2.1,<6.0.0
-cascade
+caskade
 graphviz==0.20.1
 h5py>=3.8.0
 mpmath>=1.3.0,<1.4.0
