+++ src/caustics/__init__.py
@@ -21,6 +21,7 @@ from .lenses import (
     SIE,
     SIS,
     SinglePlane,
+    BatchedPlane,
     MassSheet,
     TNFW,
     Multipole,
@@ -61,6 +62,7 @@ __all__ = [
     "SIE",
     "SIS",
     "SinglePlane",
+    "BatchedPlane",
     "MassSheet",
     "TNFW",
     "Multipole",
