### 2024-10-24 c236e89
less hacky batched lens model
