### 2024-10-25 de18beb
most editor comments addressed
