+++ paper.md
@@ -120,28 +120,28 @@ Several other simulation packages for strong gravitational lensing are already
 publicly available. The well-established `lenstronomy` package has been in use
 since 2018 [@lenstronomy]; `GLAMER` is a C++-based code for modelling complex
 and large dynamic range fields [@GLAMER]; `PyAutoLens` is also widely used
-[@PyAutoLens]; GIGA-Lens is a specialized JAX [@JAX] based gravitational lensing
-package [@GIGALens]; and Herculens is a more general JAX based lensing simulator
-package [@Herculens]; among others [GRAVLENS @Keeton2011; LENSTOOL
-@Kneib2011; SLITRONOMY @Galan2021; paltax @Wagner2024]. There are also several "in
-house" codes developed for specialized analysis which are then not publicly released
-[e.g. @Suyu2010]. `Caustics` development has been primarily focused on three
-aspects: processing speed, user experience, and flexibility. The code is
+[@PyAutoLens]; `GIGA-Lens` is a specialized JAX [@JAX] based gravitational
+lensing package [@GIGALens]; and `Herculens` is a more general JAX based lensing
+simulator package [@Herculens]; among others [`GRAVLENS` @Keeton2011; `LENSTOOL`
+@Kneib2011; `SLITRONOMY` @Galan2021; `paltax` @Wagner2024]. There are also several
+in-house codes developed for specialized analysis which are then not publicly released
+[e.g. @Suyu2010]. The development of `caustics` has been primarily focused on
+three aspects: processing speed, user experience, and flexibility. The code is
 optimized to fully exploit PyTorch's capabilities, significantly enhancing
 processing speed. The user experience is streamlined by providing three
 interfaces to the code: configuration file, object-oriented, and functional,
 where each interface level requires more expertise but allows more capabilities.
 In this way, users with all levels of gravitational lensing simulation
-experience may effectively engage with the software. Flexibility is achieved by
+experience can effectively engage with the software. Flexibility is achieved by
 a determined focus on minimalism in the core functionality of `caustics`. All of
 these elements combine to make `caustics` a capable lensing simulator to support
 machine learning applications, and classical analysis.
 
-`Caustics` fulfills a timely need for a differentiable lensing simulator.
-Several other fields have already benefitted from such simulators, for example:
+`Caustics` fills a timely need for a differentiable lensing simulator. Several
+other fields have already benefited from such simulators, for example:
 gravitational wave analysis [@Coogan2022; @Edwards2023; @Wong2023]; astronomical
 image photometry [@Stone2023]; point spread function modelling [@Desdoigts2023];
-light curves [@Millon2024]; and even generic optimization for scientific
+time series analysis [@Millon2024]; and even generic optimization for scientific
 problems [@Nikolic2018]. `Caustics` is built on lessons from other
 differentiable codes, with the goal of enabling machine learning techniques in
 the field of strong lensing. With `caustics` it will now be possible to analyze
@@ -151,31 +151,31 @@ over 100,000 lenses in a timely manner [@Hezaveh2017; @Perreault2017].
 
 `Caustics` is a gravitational lensing simulator. The purpose of the project is
 to streamline the simulation of strong gravitational lensing effects on the
-light of a background source. Primarily the focus is on all transformations
+light of a background source. The primary focus is on all transformations
 between the source plane(s) and the image plane through the lensing plane(s).
 There is minimal effort on modelling the observational elements of atmosphere,
-telescope optics, detector effects with most of these left to the user. A
+telescope optics, detector effects with most of these left to the users. A
 variety of parametric lensing profiles are included, such as: Singular
 Isothermal Ellipsoid (SIE), Elliptical Power Law (EPL), Pseudo-Jaffe,
[…]
