+++ paper.bib
@@ -27,7 +27,7 @@ journal = {Journal of Open Source Software}
 @ARTICLE{Herculens,
        author = {{Galan}, A. and {Vernardos}, G. and {Peel}, A. and {Courbin}, F. and {Starck}, J. -L.},
         title = "{Using wavelets to capture deviations from smoothness in galaxy-scale strong lenses}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {galaxies: structure, dark matter, methods: data analysis, gravitation, gravitational lensing: strong, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = dec,
@@ -45,7 +45,7 @@ archivePrefix = {arXiv},
 @ARTICLE{GIGALens,
        author = {{Gu}, A. and {Huang}, X. and {Sheu}, W. and {Aldering}, G. and {Bolton}, A.~S. and {Boone}, K. and {Dey}, A. and {Filipp}, A. and {Jullo}, E. and {Perlmutter}, S. and {Rubin}, D. and {Schlafly}, E.~F. and {Schlegel}, D.~J. and {Shu}, Y. and {Suyu}, S.~H.},
         title = "{GIGA-Lens: Fast Bayesian Inference for Strong Gravitational Lens Modeling}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {GPU computing, Markov chain Monte Carlo, Strong gravitational lensing, Einstein rings, Bayesian statistics, Cosmology, 1969, 1889, 1643, 451, 1900, 343, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Statistics - Applications},
          year = 2022,
         month = aug,
@@ -92,7 +92,7 @@ archivePrefix = {arXiv},
 @ARTICLE{GLAMER,
        author = {{Metcalf}, R. Benton and {Petkova}, Margarita},
         title = "{GLAMER - I. A code for gravitational lensing simulations with adaptive mesh refinement}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {gravitational lensing: strong, gravitational lensing: weak, gravitational lensing: micro, methods: numerical, cosmology: miscellaneous, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2014,
         month = dec,
@@ -110,7 +110,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Galan2021,
        author = {{Galan}, A. and {Peel}, A. and {Joseph}, R. and {Courbin}, F. and {Starck}, J. -L.},
         title = "{SLITRONOMY: Towards a fully wavelet-based strong lensing inversion technique}",
-      journal = {\aap},
+      journal = {Astronomy and Astrophysics},
      keywords = {gravitation, gravitational lensing: strong, methods: data analysis, techniques: image processing, galaxies: high-redshift, galaxies: structure, Astrophysics - Astrophysics of Galaxies, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2021,
         month = mar,
@@ -128,7 +128,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Suyu2010,
        author = {{Suyu}, S.~H. and {Halkola}, A.},
         title = "{The halos of satellite galaxies: the companion of the massive elliptical lens SL2S J08544-0121}",
-      journal = {\aap},
[…]
