+++ paper.md
@@ -17,7 +17,7 @@ authors:
     orcid: 0000-0001-8806-7936
     equal-contrib: true
     affiliation: "1, 2, 3"
-  - name: Adam Coogan[^2]
+  - name: Adam Coogan[^adam]
     orcid: 0000-0002-0055-1780
     equal-contrib: true
     affiliation: "1, 2, 3"
@@ -71,7 +71,7 @@ date: 19 March 2024
 bibliography: paper.bib
 ---
 
-[^2]: Work done while at UdeM, Ciela, and Mila
+[^adam]: Work done while at UdeM, Ciela, and Mila
 
 # Summary
 
@@ -239,7 +239,7 @@ a strong performance when using the batched GPU setting, especially in the more
 realistic scenario with extra computations in the simulator including 4x
 oversampling of the raytracing and the PSF convolution.
 
-![Runtime comparisons for a simple lensing setup. We compare the amount of time taken (y-axis) to generate a certain number of lensing realizations (x-axis) where a Sérsic model is lensed by an SIE mass distribution. For CPU calculations we use `Intel Gold 6148 Skylake` and for the GPU we use a `NVIDIA V100`, all tests were done at 64 bit precision. On the left, the lensing system is sampled 128 pixel resolution only at pixel midpoints. On the right, a more realistic simulation includes upsampled pixels and PSF convolution. From the two tests we see varying performance enhancements from compiled, unbatched, batched, multi-threaded, and GPU processing setups.\label{fig:runtime}](media/runtime_comparison_img.png)
+![Runtime comparisons for a simple lensing setup. We compare the amount of time taken (y-axis) to generate a certain number of lensing realizations (x-axis) where a Sérsic model is lensed by an SIE mass distribution. For CPU calculations we use `Intel Gold 6148 Skylake` and for the GPU we use a `NVIDIA V100`, all tests were done at 64 bit precision. On the left, the lensing system is sampled 128 pixel resolution only at pixel midpoints. On the right, a more realistic simulation includes upsampled pixels and PSF convolution. From the two tests we see varying performance enhancements from compiled, unbatched, batched, multi-threaded, and GPU processing setups.\label{fig:runtime}](media/runtime_comparison.png)
 
 Comparing the "caustics unbatched CPU" and "caustics batched CPU" lines we see
 that batching can provide more efficient use of the same, single CPU,
