### 2024-10-31 afe9d46
attempt footnote
