### 2024-11-01 5a66f69
docs: match citation to paper
