+++ CITATION.cff
@@ -7,29 +7,37 @@ authors:
 - family-names: "Adam"
   given-names: "Alexandre"
   orcid: "https://orcid.org/0000-0001-8806-7936"
+- family-names: "Coogan"
+  given-names: "Adam"
+  orcid: "https://orcid.org/0000-0002-0055-1780"
+- family-names: "Yantovski-Barth"
+  given-names: "Misha"
+  orcid: "https://orcid.org/0000-0001-5200-4095"
+- family-names: "Filipp"
+  given-names: "Andreas"
+  orcid: "https://orcid.org/0000-0003-4701-3469"
+- family-names: "Setiawan"
+  given-names: "Landung"
+  orcid: "https://orcid.org/0000-0002-1624-2667"
+- family-names: "Core"
+  given-names: "Cordero"
+  orcid: "https://orcid.org/0000-0002-3531-3221"
+- family-names: "Legin"
+  given-names: "Ronan"
+  orcid: "https://orcid.org/0000-0001-9459-6316"
+- family-names: "Wilson"
+  given-names: "Charles"
+  orcid: "https://orcid.org/0000-0001-7071-5528"
+- family-names: "Missael Barco"
+  given-names: "Gabriel"
+  orcid: "https://orcid.org/0009-0008-5839-5937"
 - family-names: "Perreault-Levasseur"
   given-names: "Laurence"
   orcid: "https://orcid.org/0000-0003-3544-3939"
 - family-names: "Hezaveh"
   given-names: "Yashar"
   orcid: "https://orcid.org/0000-0002-8669-5733"
-- family-names: "Core"
-  given-names: "Cordero"
-  orcid: "https://orcid.org/0000-0002-3531-3221"
-- family-names: "Setiawan"
-  given-names: "Landung"
-  orcid: "https://orcid.org/0000-0002-1624-2667"
-- family-names: "Coogan"
-  given-names: "Adam"
-- family-names: "Filipp"
-  given-names: "Andreas"
-- family-names: "Wilson"
-  given-names: "Charles"
-- family-names: "Yantovski-Barth"
-  given-names: "Misha"
-- family-names: "Legin"
-  given-names: "Ronan"
 title: "caustics"
 doi: 10.5281/zenodo.10806382
-abstract: "The lensing pipeline of the future: GPU-accelerated, automatically-differentiable, highly modular. Currently under heavy development: expect interface changes and some imprecise/untested calculations."
+abstract: "The lensing pipeline of the future: GPU-accelerated, automatically-differentiable, highly modular. Currently under heavy development."
 repository-code: "https://github.com/Ciela-Institute/caustics"
