### 2024-11-01 ff09a19
Update citation note
