### 2024-11-06 36bbddd
add chunk_size option in batched plane
