### 2024-11-06 76c125a
add shear function to base lens
