### 2024-11-06 7948c54
add shear tests vs lenstronomy
