### 2024-11-06 d1d9a48
add consistency test between fft and conv2d for LensSource
