### 2024-11-07 7d0fb06
feat: add shear calculation for all lenses (#280)
* add shear function to base lens

* add shear tests vs lenstronomy
