### 2024-11-07 9f8c176
minor changes from comments. better batch plane, mass sheet kappa
