### 2024-11-07 #7081 @ConnorStoneAstro
@ivastar I have addressed all of your editorial comments. Only ones which caused some trouble. For the footnote for co-author Adam Coogan I wasn't able to make a footnote with a different marker, so I put it at the end of the affiliations where it seemed to work. And for the LSST acronym it seems there are two versions going around. I went to the LSST website "about" page (https://www.lsst.org/about) and it seems that `Legacy Survey of Space and Time` is what they go with, so that's what I put in the paper.

As for the additional author tasks:
-  I have checked the author list and affiliations including orcid
- The latest release of caustics with all the updates is version 0.12.0
- We have a new zenodo archive for version 0.12.0. The zenodo doi link is: https://doi.org/10.5281/zenodo.14052926
- The zenodo authors and orcids match the paper
- the licenses are the same

Let me know if there's anything else I need to address.
