### 2024-11-08 085429d
jacobian of deflection angle now far more memory efficient at small compute cost
