### 2024-11-08 09d5ed4
make tests pass for NFW TNFW
