### 2024-11-13 #7081 @ivastar
@ConnorStoneAstro this is caused by the † symbol, our compilation framework really doesn't like that. Can we maybe try a footnote or a letter or number 8 instead?
