### 2024-11-13 #7081 @ConnorStoneAstro
@ivastar I never got the footnote to work, a number 8 might be a bit confusing, let's try a letter `a` and see if that works. I've uploaded the new version on the JOSS branch
