### 2024-11-13 #7081 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
