### 2024-11-13 #7081 @ivastar
@ConnorStoneAstro the EiC will do the final publication magic.

@andigu, @Jammy2211 sincere thanks for your work on this manuscript and your prompt review and responses!
