### 2024-11-15 #7081 @warrickball
Hi @ConnorStoneAstro, I'm the AEiC for this JOSS "track". Before anything else, I need to flag that the paper is currently too long. Specifically, from the [JOSS docs](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) (emphasis mine),

> The paper should be between **250-1000 words**. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

I apologise that we should've flagged this earlier and understand we might not get all the way down to 1000 words but the article is currently around 3600 words. Note the short list of items that need to be in the paper, in the same docs linked above. At a glance, I'll recommend the following:

* The "Performance" section can go, though you may wish to mention the performance relative to other packages (like lenstronomy) and the GPU performance in a few sentences under statement of need.
* The "User experience" section can go entirely. Our docs specifically say

> Given this format, a “full length” paper is not permitted, and software documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

* Similarly, the "Flexibility", "Machine Learning" and "Conclusions" don't warrant whole sections. Machine learning is already mentioned sufficiently in the "Scope" section. Some of the "Conclusions" text might be consolidated into the summary.

You're welcome to transplant any of this material to your docs. Let me know if you have any questions! If you'd like to run the word count again, you can comment with `@editorialbot check repository`. I fully acknowledge that words counts are a bit uncertain but I'd at least like to get this under 2000 words, if not closer to 1000.
