### 2024-11-15 2c5eed2
Reduce paper size to JOSS requirements
