+++ paper.md
@@ -112,12 +112,14 @@ computational acceleration via GPUs and/or algorithmic advances such as
 automatic differentiation are needed to reduce the analysis timescales. Machine
 learning will be critical to achieve the necessary speed to process these
 lenses, it will also be needed to meet the complexity of strong lens modelling.
-`caustics` is built with the future of lensing in mind, using `PyTorch`
-[@pytorch] to accelerate the low-level computation and enable deep learning
-algorithms which rely on automatic differentiation. Automatic differentiation
-also benefits classical algorithms such as Hamiltonian Monte Carlo [@hmc]. With
-these tools available, `caustics` provides greater than two orders of magnitude
-acceleration to most standard operations, enabling previously impractical
+Literature on machine learning applications in strong gravitational lensing
+underscores this need [@Brehmer2019; @Chianese2020; @Coogan2020;
+@Mishra2022; @Karchev2022; @Karchev2022b]. `caustics` is built with the future of
+lensing in mind, using `PyTorch` [@pytorch] to accelerate the low-level computation
+and enable deep learning algorithms which rely on automatic differentiation. Automatic
+differentiation also benefits classical algorithms such as Hamiltonian Monte Carlo
+[@hmc]. With these tools available, `caustics` provides greater than two orders of
+magnitude acceleration to most standard operations, enabling previously impractical
 analyses at scale.
 
 Several other simulation packages for strong gravitational lensing are already
@@ -132,14 +134,16 @@ in-house codes developed for specialized analysis which are then not publicly re
 [e.g. @Suyu2010]. The development of `caustics` has been primarily focused on
 three aspects: processing speed, user experience, and flexibility. The code is
 optimized to fully exploit PyTorch's capabilities, significantly enhancing
-processing speed. The user experience is streamlined by providing three
-interfaces to the code: configuration file, object-oriented, and functional,
-where each interface level requires more expertise but allows more capabilities.
-In this way, users with all levels of gravitational lensing simulation
-experience can effectively engage with the software. Flexibility is achieved by
-a determined focus on minimalism in the core functionality of `caustics`. All of
-these elements combine to make `caustics` a capable lensing simulator to support
-machine learning applications, and classical analysis.
+processing speed. Runtimes are comparable to the widely used `lenstronomy` when
+on CPU, and can be over 1000 times faster on GPU depending on configuration. The
+user experience is streamlined by providing three interfaces to the code:
+configuration file, object-oriented, and functional, where each interface level
+requires more expertise but allows more capabilities. In this way, users with
+all levels of gravitational lensing simulation experience can effectively engage
+with the software. Flexibility is achieved by a determined focus on minimalism
+in the core functionality of `caustics`. All of these elements combine to make
+`caustics` a capable lensing simulator to support machine learning applications,
+and classical analysis.
 
 `Caustics` fills a timely need for a differentiable lensing simulator. Several
 other fields have already benefited from such simulators, for example:
@@ -170,32 +174,32 @@ lists using templates provided in the documentation.
 Once a lensing system has been defined, `caustics` can then perform various
 computational operations on the system such as raytracing through the lensing
 system, forwards and backwards. Users can compute the lensing potential,
-convergence, deflection field, time delay field, and magnification. All of these
-operations can readily be performed in a multi-plane setting to account for
-interlopers or multiple sources. Because the code is differentiable (via
+convergence, deflection field, time delay field, shear, and magnification. All
+of these operations can readily be performed in a multi-plane setting to account
+for interlopers or multiple sources. Because the code is differentiable (via
 PyTorch), one can easily also compute the derivatives of these quantities, such
[…]
