+++ paper.md
@@ -80,24 +80,27 @@ bibliography: paper.bib
 # Summary
 
 Gravitational lensing is the deflection of light rays due to the gravity of
-intervening masses. This phenomenon is observed at a variety of scales and
-configurations, involving any non-uniform mass such as planets, stars, galaxies,
-clusters of galaxies, and even the large-scale structure of the Universe. Strong
-lensing occurs when the distortions are significant and multiple images of the
-background source are observed. The lens and lensed object(s) must be aligned on
-the sky within $\sim 1$ arcsecond for galaxy-galaxy lensing, or tens of
-arcseconds for cluster-galaxy lensing. As the discovery of lens systems has
-grown to the low thousands, these systems have become pivotal for precision
-measurements and addressing critical questions in astrophysics. Notably, they
-are a tool for understanding a wide range of astrophysical phenomena including:
-dark matter [e.g. @Hezaveh2016; @Vegetti2014], supernovae [e.g. @Rodney2021], quasars
-[e.g. @Peng2006], the first stars [e.g. @Welch2022], and the Universe's expansion
-rate [e.g. @holycow]. With future surveys expected to discover hundreds of
-thousands of lensing systems, the modelling and simulation of such systems must
-be done at orders of magnitude larger scale than ever before. Here we present
-`caustics`, a Python package designed to facilitate machine learning and
-Bayesian methods to handle the extensive computational demands of modelling such
-a vast number of lensing systems.
+intervening masses. This phenomenon is observed at a variety of configurations,
+involving any non-uniform mass such as planets, stars, galaxies, clusters of
+galaxies, and even the large-scale structure of the Universe. Strong lensing
+occurs when the distortions are significant and multiple images of the
+background source are observed. The lens and lensed object(s) must be aligned
+closely on the sky. As the discovery of lens systems has grown to the low
+thousands, these systems have become pivotal for precision measurements in
+astrophysics. Notably, for phenomena including: dark matter [e.g.
+@Hezaveh2016; @Vegetti2014], supernovae [e.g. @Rodney2021], quasars [e.g.
+@Peng2006],
+the first stars [e.g. @Welch2022], and the Universe's expansion rate [e.g.
+@holycow].
+With future surveys expected to discover hundreds of thousands of lensing
+systems, the modelling and simulation of such systems must be done at orders of
+magnitude larger scale than ever before. Here we present `caustics`, a Python
+package designed to facilitate machine learning and Bayesian methods to handle
+the extensive computational demands of modelling such a vast number of lensing
+systems.
+
+<!-- within $\sim 1$ arcsecond for galaxy-galaxy lensing, or tens of
+arcseconds for cluster-galaxy lensing -->
 
 # Statement of need
 
@@ -106,21 +109,22 @@ and Time, the Roman Core Community Surveys, and the Euclid wide survey, are
 expected to uncover hundreds of thousands of gravitational lenses
 [@Collett2015], dramatically increasing the scientific potential of
 gravitational lensing studies. Currently, analyzing a single lensing system can
-take several days or weeks, which will be infeasible as the number of known
-lenses increases by orders of magnitude. Thus, advancements such as
-computational acceleration via GPUs and/or algorithmic advances such as
-automatic differentiation are needed to reduce the analysis timescales. Machine
-learning will be critical to achieve the necessary speed to process these
-lenses, it will also be needed to meet the complexity of strong lens modelling.
-Literature on machine learning applications in strong gravitational lensing
-underscores this need [@Brehmer2019; @Chianese2020; @Coogan2020;
-@Mishra2022; @Karchev2022; @Karchev2022b]. `caustics` is built with the future of
-lensing in mind, using `PyTorch` [@pytorch] to accelerate the low-level computation
-and enable deep learning algorithms which rely on automatic differentiation. Automatic
[…]
