### 2024-11-15 598c19e
add back runtime comparison
