### 2024-11-15 #7081 @ConnorStoneAstro
@warrickball ah, my apologies for not identifying this sooner. I had no simple word counter for markdown so I just assumed the bot checks would flag such an issue, perhaps this could be added to avoid future mishaps? I should also ask if an appendix section is allowed? I assume not, but there is no information either way on the JOSS docs.

I have cut the paper down to approximately 1000 words (manually excluding acknowledgments and meta-text), but I will require a re-confirmation with all the co-authors before moving forward. I'll notify you when it is ready.
