### 2024-11-20 6cff853
lenssource now handles larger grids
