### 2024-11-20 #7081 @warrickball
That's great, thanks!

> ... I just assumed the bot checks would flag such an issue, perhaps this could be added to avoid future mishaps?

Absolutely. We're working on trying to make the automatic messages more obviously indicate when a submission is over the ~1000 word limit.

> I should also ask if an appendix section is allowed?

No, though you're right that we could make this explicit in the docs. The paper is intended to principally motivate why the software exists, though this will itself involve describing some background and features. We expect most other details to go into the software documentation, which is also reviewed.
