### 2024-11-20 #7081 @warrickball
Aha, I realise the word count still counts everything that's been commented, which isn't a problem.
