+++ paper.md
@@ -99,9 +99,6 @@ package designed to facilitate machine learning and Bayesian methods to handle
 the extensive computational demands of modelling such a vast number of lensing
 systems.
 
-<!-- within $\sim 1$ arcsecond for galaxy-galaxy lensing, or tens of
-arcseconds for cluster-galaxy lensing -->
-
 # Statement of need
 
 The next generation of astronomical surveys, such as the Legacy Survey of Space
@@ -119,12 +116,6 @@ applications in strong gravitational lensing underscores this need
 @Karchev2022b]. `caustics` is built with the future of lensing in mind, using `PyTorch`
 [@pytorch] to accelerate the low-level computation and enable deep learning algorithms
 which rely on automatic differentiation.
-
-<!-- Automatic
-differentiation also benefits classical algorithms such as Hamiltonian Monte Carlo
-[@hmc]. With these tools available, `caustics` provides greater than two orders of
-magnitude acceleration to most standard operations, enabling previously impractical
-analyses at scale. -->
 
 Several other simulation packages for strong gravitational lensing are already
 publicly available. The well-established `lenstronomy` package has been in use
@@ -146,16 +137,6 @@ functionality of `caustics` and encouraging user extension.
 
 ![Runtime comparisons for a simple lensing setup. We compare the amount of time taken (y-axis) to generate a certain number of lensing realizations (x-axis) where a Sérsic model is lensed by an SIE mass distribution. For CPU calculations we use `Intel Gold 6148 Skylake` and for the GPU we use a `NVIDIA V100`, all tests were done at 64 bit precision. On the left, the lensing system is sampled 128 pixel resolution only at pixel midpoints. On the right, a more realistic simulation includes upsampled pixels and PSF convolution. From the two tests we see varying performance enhancements from compiled, unbatched, batched, multi-threaded, and GPU processing setups.\label{fig:runtime}](media/runtime_comparison.png)
 
-<!--
-, where each interface level
-requires more expertise but allows more capabilities. In this way, users with
-all levels of gravitational lensing simulation experience can effectively engage
-with the software.
-
-All of these elements combine to make
-`caustics` a capable lensing simulator to support machine learning applications,
-and classical analysis. -->
-
 `Caustics` fills a timely need for a differentiable lensing simulator. Several
 other fields have already benefited from such simulators, for example:
 gravitational wave analysis [@Coogan2022; @Edwards2023; @Wong2023]; astronomical
@@ -163,10 +144,6 @@ image photometry [@Stone2023]; point spread function modelling [@Desdoigts2023];
 time series analysis [@Millon2024]; and even generic optimization for scientific
 problems [@Nikolic2018]. With `caustics` it will now be possible to analyze over
 100,000 lenses in a timely manner [@Hezaveh2017; @Perreault2017].
-
- <!-- `Caustics` is built on lessons from other
-differentiable codes, with the goal of enabling machine learning techniques in
-the field of strong lensing.  -->
 
 # Scope
 
@@ -181,21 +158,12 @@ Pseudo-Jaffe, Navarro-Frenk-White (NFW), and External Shear. Additionally, it
 offers non-parametric representations such as a gridded convergence or a
 potential field and pixelized sources.
 
-<!-- For the background source `caustics` provides a Sérsic light profile, as
-well as a pixelized light image. Users can easily extend these lens and source
-lists using templates provided in the documentation. -->
-
 Once a lensing system has been defined, `caustics` can then perform various
 computational operations on the system such as raytracing through the lensing
 system, forwards and backwards. Users can compute the lensing potential,
 convergence, deflection field, time delay field, shear, and magnification. All
 of these operations can readily be performed in a multi-plane setting to account
[…]
