### 2024-11-20 #7081 @ConnorStoneAstro
@warrickball I have pushed a new version which removes the commented text. My manual word counting of each section seems to land around 1050 words, I suspect the bot will get 1500 words if it is including the acknowledgements and metadata.
