### 2024-11-21 4a091ca
add padding mode for pixelated light source
