### 2024-11-21 b0e3c0d
Guassian to torch
