### 2024-11-21 e363304
refactor: Change Parametrized system over to caskade (#275)
* first pass convert to caskade

* clear pydantic, some tests pass

* all tests pass

* params by args

* fix requirements

* add param metadata

* get tutorials to work

* adding batched plane

* get tutorials to run

* fix qso fit example

* less hacky batched lens model

* with math expressions

* add chunk_size option in batched plane

* shift order of parameters in lenssource

* minor changes from comments. better batch plane, mass sheet kappa

* jacobian of deflection angle now far more memory efficient at small compute cost

* 
[…]
