### 2024-11-21 50b0c84
docs: Minor corrections
