### 2024-11-21 7e5782a
fix citation.cff title
