### 2024-11-21 #7081 @ConnorStoneAstro
@warrickball ok, the new trimmed version of the paper is ready and approved by the co-authors. We have also set the version number for caustics to `v1.0.0`. The new zenodo DOI for this version is `10.5281/zenodo.14201027` (the general DOI for all versions is `10.5281/zenodo.10806382`) perhaps that could be used as the JOSS publication version? Please let me know if anything seems amiss
