### 2024-11-22 b34d936
Editorial tweaks for JOSS publication
