+++ paper.bib
@@ -27,7 +27,7 @@ journal = {Journal of Open Source Software}
 @ARTICLE{Herculens,
        author = {{Galan}, A. and {Vernardos}, G. and {Peel}, A. and {Courbin}, F. and {Starck}, J. -L.},
         title = "{Using wavelets to capture deviations from smoothness in galaxy-scale strong lenses}",
-      journal = {Astronomy and Astrophysics},
+      journal = {Astronomy \& Astrophysics},
      keywords = {galaxies: structure, dark matter, methods: data analysis, gravitation, gravitational lensing: strong, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = dec,
@@ -44,7 +44,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{GIGALens,
        author = {{Gu}, A. and {Huang}, X. and {Sheu}, W. and {Aldering}, G. and {Bolton}, A.~S. and {Boone}, K. and {Dey}, A. and {Filipp}, A. and {Jullo}, E. and {Perlmutter}, S. and {Rubin}, D. and {Schlafly}, E.~F. and {Schlegel}, D.~J. and {Shu}, Y. and {Suyu}, S.~H.},
-        title = "{GIGA-Lens: Fast Bayesian Inference for Strong Gravitational Lens Modeling}",
+        title = {{GIGA-Lens}: Fast {Bayesian} Inference for Strong Gravitational Lens Modeling},
       journal = {The Astrophysical Journal},
      keywords = {GPU computing, Markov chain Monte Carlo, Strong gravitational lensing, Einstein rings, Bayesian statistics, Cosmology, 1969, 1889, 1643, 451, 1900, 343, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Statistics - Applications},
          year = 2022,
@@ -63,7 +63,7 @@ archivePrefix = {arXiv},
 
 @software{Kneib2011,
        author = {{Kneib}, Jean-Paul and {Bonnet}, Henri and {Golse}, Ghyslain and {Sand}, David and {Jullo}, Eric and {Marshall}, Phil},
-        title = "{LENSTOOL: A Gravitational Lensing Software for Modeling Mass Distribution of Galaxies and Clusters (strong and weak regime)}",
+        title = {{LENSTOOL}: A Gravitational Lensing Software for Modeling Mass Distribution of Galaxies and Clusters (strong and weak regime)},
  howpublished = {Astrophysics Source Code Library, record ascl:1102.004},
          year = 2011,
         month = feb,
@@ -74,25 +74,27 @@ archivePrefix = {arXiv},
 
 @ARTICLE{Wagner2024,
        author = {{Wagner-Carena}, Sebastian and {Lee}, Jaehoon and {Pennington}, Jeffrey and {Aalbers}, Jelle and {Birrer}, Simon and {Wechsler}, Risa H.},
-        title = "{A Strong Gravitational Lens Is Worth a Thousand Dark Matter Halos: Inference on Small-Scale Structure Using Sequential Methods}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+        title = {A Strong Gravitational Lens Is Worth a Thousand Dark Matter Halos: Inference on Small-scale Structure Using Sequential Methods},
+      journal = {The Astrophysical Journal},
+     keywords = {Dark matter, Cosmology, Strong gravitational lensing, 353, 343, 1643, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2024,
-        month = apr,
-          eid = {arXiv:2404.14487},
[…]
