+++ paper.md
@@ -87,7 +87,7 @@ occurs when the distortions are significant and multiple images of the
 background source are observed. The lens and lensed object(s) must be aligned
 closely on the sky. As the discovery of lens systems has grown to the low
 thousands, these systems have become pivotal for precision measurements in
-astrophysics. Notably, for phenomena including: dark matter [e.g.
+astrophysics, notably for phenomena including dark matter [e.g.
 @Hezaveh2016; @Vegetti2014], supernovae [e.g. @Rodney2021], quasars [e.g.
 @Peng2006],
 the first stars [e.g. @Welch2022], and the Universe's expansion rate [e.g.
@@ -109,20 +109,20 @@ gravitational lensing studies. Currently, analyzing a single lensing system can
 take several days or weeks, which will soon be infeasible to scale. Thus,
 advancements such as GPU acceleration and/or automatic differentiation are
 needed to reduce the analysis timescales. Machine learning will be critical to
-achieve the necessary speed to process these lenses, it will also be needed to
+achieve the necessary speed to process these lenses. It will also be needed to
 meet the complexity of strong lens modelling. Literature on machine learning
 applications in strong gravitational lensing underscores this need
 [@Brehmer2019; @Chianese2020; @Coogan2020; @Mishra2022; @Karchev2022;
 @Karchev2022b]. `caustics` is built with the future of lensing in mind, using `PyTorch`
-[@pytorch] to accelerate the low-level computation and enable deep learning algorithms
+[@pytorch] to accelerate the low-level computation and enable deep learning algorithms,
 which rely on automatic differentiation.
 
 Several other simulation packages for strong gravitational lensing are already
 publicly available. The well-established `lenstronomy` package has been in use
 since 2018 [@lenstronomy]; `GLAMER` is a C++-based code for modelling complex
 and large dynamic range fields [@GLAMER]; `PyAutoLens` is also widely used
-[@PyAutoLens]; `GIGA-Lens` is a specialized JAX [@JAX] based gravitational
-lensing package [@GIGALens]; and `Herculens` is a more general JAX based lensing
+[@PyAutoLens]; `GIGA-Lens` is a specialized JAX-based [@JAX] gravitational
+lensing package [@GIGALens]; and `Herculens` is a more general JAX-based lensing
 simulator package [@Herculens]; among others [`GRAVLENS` @Keeton2011; `LENSTOOL`
 @Kneib2011; `SLITRONOMY` @Galan2021; `paltax` @Wagner2024]. There are also several
 in-house codes developed for specialized analysis which are then not publicly released
@@ -135,23 +135,23 @@ interfaces to the code: configuration file, object-oriented, and functional.
 Flexibility is achieved by a determined focus on minimalism in the core
 functionality of `caustics` and encouraging user extension.
 
-![Runtime comparisons for a simple lensing setup. We compare the amount of time taken (y-axis) to generate a certain number of lensing realizations (x-axis) where a Sérsic model is lensed by an SIE mass distribution. For CPU calculations we use `Intel Gold 6148 Skylake` and for the GPU we use a `NVIDIA V100`, all tests were done at 64 bit precision. On the left, the lensing system is sampled 128 pixel resolution only at pixel midpoints. On the right, a more realistic simulation includes upsampled pixels and PSF convolution. From the two tests we see varying performance enhancements from compiled, unbatched, batched, multi-threaded, and GPU processing setups.\label{fig:runtime}](media/runtime_comparison.png)
[…]
