+++ paper.bib
@@ -352,7 +352,7 @@ archivePrefix = {arXiv},
 }
 
 @ARTICLE{Millon2024,
-       author = {{Millon}, Martin and {Michalewicz}, Kevin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric and {Marshall}, Philip J.},
+       author = {{Million}, Martin and {Michalewicz}, Kevin and {Dux}, Fr{\'e}d{\'e}ric and {Courbin}, Fr{\'e}d{\'e}ric and {Marshall}, Philip J.},
         title = {Image Deconvolution and Point-spread Function Reconstruction with {STARRED}: A Wavelet-based Two-channel Method Optimized for Light-curve Extraction},
       journal = {The Astronomical Journal},
      keywords = {Deconvolution, Photometry, Light curves, Astronomy image processing, 1910, 1234, 918, 2306, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Astrophysics of Galaxies},
