### 2024-11-22 #7081 @warrickball
Hi @ConnorStoneAstro — I've open [a PR](https://github.com/Ciela-Institute/caustics/pull/290) with some editorial tweaks. If you check those are all okay, then I'll give the publication pipeline another test run and then we'll probably be ready to publish your article.
