### 2024-11-22 #7081 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
